### 2023-07-14 #5655 @patquem
Hi @teuben,

Thank you very much for your comments.

Regarding your 1st comment, I did not know the FITS format.
Finding a name for an application is getting more and more complicated. For the first 4 months of the project, it was called: **spectrafit**, but when I thought of a pypi repository, that name was already in use. 2-3 months ago I changed to **fitspy** (not used in pypi and as a hint at hyperspy, a library which is a reference in my community). Since then, I have been promoting the tool with that name. I'm afraid that a new change is no longer confusing in my community.
but by any chance, do you have a relevant name to suggest to me? : )

Regarding your 2nd comment: you are absolutely right. I will add these 2 references which are relevant in the article.

(Note that I'm going on vacation tomorrow for 2 weeks. I will therefore make the corrections on my return (end of July).)
