### 2023-07-25 #5655 @Kevin-Mattheus-Moerman
@patquem thanks for this submission. I am the AEiC for this track and here to help with initial steps to to ensure this work is in scope for JOSS. On the latter point, I have just triggered a scope review by our editorial board. This is to review if this work conforms to our [substantial scholarly effort criteria](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). This scope review should be completed within about 2 weeks.
