### 2023-07-31 ca312b5
acknowledgments revisited + new references
