### 2023-07-31 #5655 @patquem
changes:
- references to specutils and pyspeckit added
- acknowledgments revisited
