+++ .github/workflows/pytest.yml
@@ -6,9 +6,6 @@ on:
   pull_request:
     branches: [ "test_ci" ]
 
-env:
-  PYTHONPATH: ${{ vars.GITHUB_WORKSPACE }}
-
 jobs:
   build:
 
@@ -31,4 +28,4 @@ jobs:
         python setup.py install
     - name: Test with pytest
       run: |
-        pytest tests/
+        python -m pytest tests/
