+++ .github/workflows/pytest.yml
@@ -9,11 +9,11 @@ on:
 jobs:
   build:
 
-    runs-on: windows-latest
+    runs-on: ubuntu-latest
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.9", "3.10"] # # scipy-1.11.1: Python version >= 3.9 required
+        python-version: ["3.9", "3.10"] # scipy-1.11.1: Python version >= 3.9 required
 
     steps:
     - uses: actions/checkout@v3
