+++ .github/workflows/pytest.yml
@@ -2,18 +2,18 @@ name: Python test
 
 on:
   push:
-    branches: [ "main" ]
+    branches: [ "test_xdisplay" ]
   pull_request:
-    branches: [ "main" ]
+    branches: [ "test_xdisplay" ]
 
 jobs:
   build:
 
-    runs-on: ubuntu-latest
+    runs-on: windows-latest
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.7", "3.8", "3.9", "3.10"]
+        python-version: ["3.9", "3.10"] # scipy-1.11.1: Python version >= 3.9 required
 
     steps:
     - uses: actions/checkout@v3
