+++ .github/workflows/pytest.yml
@@ -2,9 +2,9 @@ name: Python test
 
 on:
   push:
-    branches: [ "test_xdisplay" ]
+    branches: [ "main" ]
   pull_request:
-    branches: [ "test_xdisplay" ]
+    branches: [ "main" ]
 
 jobs:
   build:
