### 2023-08-02 195f647
scipy requires python>=3.9
