+++ .github/workflows/pytest.yml
@@ -25,7 +25,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         python -m pip install pytest
-        python setup.py install
+        pip install .
     - name: Test with pytest
       run: |
         python -m pytest tests/
