+++ .github/workflows/pytest.yml
@@ -13,7 +13,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.7"] # scipy-1.11.1: Python version >= 3.9 required
+        python-version: ["3.7", "3.8", "3.9", "3.10"] # scipy-1.11.1: Python version >= 3.9 required
 
     steps:
     - uses: actions/checkout@v3
@@ -25,7 +25,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         python -m pip install pytest
-        pip install .
+        python -m pip install .
     - name: Test with pytest
       run: |
         python -m pytest tests/
