+++ .github/workflows/pytest.yml
@@ -2,9 +2,9 @@ name: Python test
 
 on:
   push:
-    branches: [ "fix_ci_deps" ]
+    branches: [ "main" ]
   pull_request:
-    branches: [ "fix_ci_deps" ]
+    branches: [ "main" ]
 
 jobs:
   build:
