+++ tests/test_gui_2d_maps.py
@@ -2,9 +2,10 @@ import pytest
 from pytest import approx
 
 from examples.ex3_gui_2d_maps import gui_2d_maps
-from utils import extract_results
+from utils import extract_results, display_is_ok
 
 
+@pytest.mark.skipif(not display_is_ok(), reason="DISPLAY problem")
 def test_gui_2d_maps(tmp_path):
     gui_2d_maps(dirname_res=tmp_path)
 
