+++ tests/test_gui_auto_decomposition.py
@@ -2,9 +2,10 @@ import pytest
 from pytest import approx
 
 from examples.ex0_gui_auto_decomposition import gui_auto_decomposition
-from utils import extract_results
+from utils import extract_results, display_is_ok
 
 
+@pytest.mark.skipif(not display_is_ok(), reason="DISPLAY problem")
 def test_gui_auto_decomposition(tmp_path):
     gui_auto_decomposition(dirname_res=tmp_path)
 
