+++ .github/workflows/pytest.yml
@@ -2,14 +2,14 @@ name: Python test
 
 on:
   push:
-    branches: [ "main" ]
+    branches: [ "test_xdisplay" ]
   pull_request:
     branches: [ "main" ]
 
 jobs:
   build:
 
-    runs-on: windows-latest
+    runs-on: ubuntu-latest
     strategy:
       fail-fast: false
       matrix:
@@ -23,6 +23,8 @@ jobs:
         python-version: ${{ matrix.python-version }}
     - name: Install dependencies
       run: |
+        Xvfb :99 -screen 0 1024x768x16 &
+        export DISPLAY=:99
         python -m pip install --upgrade pip
         python -m pip install pytest
         python setup.py install
