+++ .github/workflows/pytest.yml
@@ -25,9 +25,10 @@ jobs:
       run: |
         Xvfb :99 -screen 0 1024x768x16 &
         export DISPLAY=:99
-        python -m pip install --upgrade pip
-        python -m pip install pytest
-        python setup.py install
-    - name: Test with pytest
-      run: |
-        python -m pytest tests/
+        ps aux | grep Xvfb
+    #     python -m pip install --upgrade pip
+    #     python -m pip install pytest
+    #     python setup.py install
+    # - name: Test with pytest
+    #   run: |
+    #     python -m pytest tests/
