+++ .github/workflows/pytest.yml
@@ -23,9 +23,9 @@ jobs:
         python-version: ${{ matrix.python-version }}
     - name: Install dependencies
       run: |
-        Xvfb :99 -screen 0 1024x768x16 &
+        xvfb :99 -screen 0 1024x768x16 &
         export DISPLAY=:99
-        ps aux | grep Xvfb
+        ps aux | grep xvfb
     #     python -m pip install --upgrade pip
     #     python -m pip install pytest
     #     python setup.py install
