+++ fitspy/display.py
@@ -1,0 +1,6 @@
+import os
+import tkinter
+print("Hello !!")
+os.environ['DISPLAY'] = ':99'
+print("test")
+tkinter.Tk()
