+++ .github/workflows/pytest.yml
@@ -23,9 +23,10 @@ jobs:
         python-version: ${{ matrix.python-version }}
     - name: Install dependencies
       run: |
-        xvfb :99 -screen 0 1024x768x16 &
+        Xvfb :99
         export DISPLAY=:99
-        ps aux | grep xvfb
+        ps aux | grep Xvfb
+        python -m display.py
     #     python -m pip install --upgrade pip
     #     python -m pip install pytest
     #     python setup.py install
