+++ .github/workflows/pytest.yml
@@ -26,10 +26,12 @@ jobs:
         Xvfb :99
         export DISPLAY=:99
         ps aux | grep Xvfb
-        python -m display.py
-    #     python -m pip install --upgrade pip
-    #     python -m pip install pytest
-    #     python setup.py install
-    # - name: Test with pytest
+        python -m pip install --upgrade pip
+        python -m pip install pytest
+        python setup.py install
     #   run: |
     #     python -m pytest tests/
+    - name: Test with pytest
+      run: |
+        python -m display.py
+
