+++ .github/workflows/pytest.yml
@@ -23,7 +23,7 @@ jobs:
         python-version: ${{ matrix.python-version }}
     - name: Install dependencies
       run: |
-        Xvfb :99
+        Xvfb :99 &
         export DISPLAY=:99
         ps aux | grep Xvfb
         python -m pip install --upgrade pip
