+++ .github/workflows/pytest.yml
@@ -33,5 +33,5 @@ jobs:
     #     python -m pytest tests/
     - name: Test with pytest
       run: |
-        python -m display
+        python display.py
 
