+++ tests/utils.py
@@ -27,4 +27,9 @@ def display_is_ok():
         tkinter.Tk()
         return True
     except tkinter.TclError:
-        return False
+        try:
+            os.environ['DISPLAY'] = ':99'  # opened port in pytest.yml
+            tkinter.Tk()
+            return True
+        except tkinter.TclError:
+            return False
