+++ .github/workflows/pytest.yml
@@ -23,9 +23,6 @@ jobs:
         python-version: ${{ matrix.python-version }}
     - name: Install dependencies
       run: |
-        Xvfb :99 &
-        export DISPLAY=:99
-        ps aux | grep Xvfb
         python -m pip install --upgrade pip
         python -m pip install pytest
         python setup.py install
@@ -33,5 +30,7 @@ jobs:
     #     python -m pytest tests/
     - name: Test with pytest
       run: |
+        Xvfb :99 &
+        export DISPLAY=:99
         python display.py
 
