+++ .github/workflows/pytest.yml
@@ -13,7 +13,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.9", "3.10"] # scipy-1.11.1: Python version >= 3.9 required
+        python-version: ["3.9"] # scipy-1.11.1: Python version >= 3.9 required
 
     steps:
     - uses: actions/checkout@v3
@@ -30,7 +30,7 @@ jobs:
     #     python -m pytest tests/
     - name: Test with pytest
       run: |
-        echo ${{ matrix.python-version }}//.
+        echo ${${{ matrix.python-version }}//.}
       # Xvfb :99 &
       # export DISPLAY=:99
       # python display.py
