### 2023-08-03 fb99252
manage dyn display
