### 2023-08-03 dc82996
manage variable DISPLAY
