+++ .github/workflows/pytest.yml
@@ -21,11 +21,11 @@ jobs:
         uses: actions/setup-python@v3
         with:
           python-version: ${{ matrix.python-version }}
-      # - name: Install dependencies
-      #   run: |
-      #     python -m pip install --upgrade pip
-      #     python -m pip install pytest
-      #     python -m pip install .
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          python -m pip install pytest
+          python -m pip install .
       - name: Test with pytest
         run: |
           PY_VERSION=${{ matrix.python-version }}
@@ -33,6 +33,5 @@ jobs:
           DISPLAY=:${PY_VERSION//.}
           export DISPLAY
           Xvfb $DISPLAY &
-          python display.py
-        # python -m pytest tests/
+          python -m pytest tests/ -vv
 
