### 2023-08-03 71acf75
print dirname
