+++ examples/ex4_inline_auto_decomposition.py
@@ -17,7 +17,7 @@ def inline_auto_decomposition(verbosity=True, show_plots=False):
 
     fnames = glob.glob(os.path.join(DATA, 'spectra_1', '*.txt'))
     dirname = os.path.join(DATA, 'spectra_1')
-    print("CHECK", dirname, os.path.isdir(dirname))
+    print("CHECK INLINE", dirname, os.path.isdir(dirname), os.path.abspath(dirname))
 
     # spectra fitting processing
     spectra_list = []
