+++ examples/ex4_inline_auto_decomposition.py
@@ -17,7 +17,8 @@ def inline_auto_decomposition(verbosity=True, show_plots=False):
 
     fnames = glob.glob(os.path.join(DATA, 'spectra_1', '*.txt'))
     dirname = os.path.join(DATA, 'spectra_1')
-    print("CHECK INLINE", dirname, os.path.isdir(dirname), os.path.abspath(dirname))
+    dirname = os.path.abspath(dirname)
+    print("CHECK INLINE", dirname, os.path.isdir(dirname))
 
     # spectra fitting processing
     spectra_list = []
