### 2023-08-03 3229e71
add DATA in .examples
