+++ examples/__init__.py
@@ -1,1 +1,3 @@
+import pathlib
 
+DATA = pathlib.Path(__file__).parent / 'data'
