+++ examples/ex0_gui_auto_decomposition.py
@@ -5,8 +5,7 @@ import os
 import tkinter as tk
 
 from fitspy.app.gui import Appli
-
-DATA = os.path.join('..', 'examples', 'data')
+from examples import DATA
 
 
 def gui_auto_decomposition(dirname_res=None):
