+++ tests/test_inline_auto_decomposition.py
@@ -22,4 +22,4 @@ def test_inline_auto_decomposition():
     for spectrum, ref in zip(spectra, refs):
         ampli = spectrum.models[0].param_hints['ampli']['value']
         assert np.sum(spectrum.y) == approx(ref[0])
-        assert ampli == approx(ref[1], rel=1e-3)
+        assert ampli == approx(ref[1], rel=1e-2)
