### 2023-08-03 44fa9a5
exclude fwhm comparison
