+++ .github/workflows/pytest.yml
@@ -2,7 +2,7 @@ name: Python test
 
 on:
   push:
-    branches: [ "main" ]
+    branches: [ "test_xdisplay2" ]
   pull_request:
     branches: [ "main" ]
 
@@ -28,9 +28,6 @@ jobs:
           python -m pip install .
       - name: Test with pytest
         run: |
-          PY_VERSION=${{ matrix.python-version }}
-          export PY_VERSION
-          DISPLAY=:${PY_VERSION//.}
-          export DISPLAY
-          Xvfb $DISPLAY &
+          export DISPLAY=:99
+          Xvfb :99 &
           python -m pytest tests/
