### 2023-08-04 6526ea2
move PATHs in fitspy/__init__.py
