+++ fitspy/__init__.py
@@ -1,0 +1,5 @@
+import pathlib
+
+ROOT = pathlib.Path(__file__).parent.parent
+EXAMPLES = ROOT / 'examples'
+DATA = EXAMPLES / 'data'
