+++ examples/ex4_inline_auto_decomposition.py
@@ -9,7 +9,7 @@ from lmfit import report_fit
 
 from fitspy.spectra import Spectrum, Spectra
 from fitspy.utils import hsorted
-from examples import DATA
+from fitspy import DATA
 
 
 def inline_auto_decomposition(verbosity=True, show_plots=False):
