+++ examples/__init__.py
@@ -1,3 +1,0 @@
-import pathlib
-
-DATA = pathlib.Path(__file__).parent / 'data'
