### 2023-08-31 f1df5a8
remove derivatives info in docstring
