### 2023-09-01 e14fccd
reintroduce find_packages()
