+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = fitspy
-version = 2023.4
+version = 2023.5
 maintainer = Patrick QUEMERE
 maintainer_email = patrick.quemere@cea.fr
 description = Generic tool to fit spectra in python
