### 2023-09-05 #5655 @phibeck
:wave: @tillsten, @maurov, & @ivergara, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
