### 2023-09-06 abb71b4
Fix for packaging / manifest
