+++ setup.cfg
@@ -1,35 +1,0 @@
-[metadata]
-name = fitspy
-version = 2023.5
-maintainer = Patrick QUEMERE
-maintainer_email = patrick.quemere@cea.fr
-description = Generic tool to fit spectra in python
-long_description = file: README.md
-long_description_content_type = text/markdown
-url = https://github.com/CEA-MetroCarac/fitspy
-classifiers =
-    Programming Language :: Python :: 3
-    License :: OSI Approved :: GNU General Public License v3 (GPLv3)
-    Operating System :: Windows
-    Operating System :: POSIX :: Linux
-    Operating System :: MacOS :: MacOS X
-
-[options]
-zip_safe = False
-packages = ['fitspy']
-include_package_data = False
-python_requires = >=3.7
-install_requires =
-    numpy
-    pandas
-    matplotlib
-    scipy
-    lmfit
-    pywin32; platform_system == "Windows"
-
-[options.packages.find]
-where = fitspy
-
-[options.entry_points]
-gui_scripts =
-    fitspy = fitspy.app.gui:fitspy_launcher
