+++ MANIFEST.in
@@ -1,1 +1,4 @@
-exclude docs examples
+graft examples
+graft doc
+graft tests
+global-exclude *.pyc __pycache__
