### 2023-09-06 47fa272
Fix for packaging / manifest
