### 2023-09-08 #5655 @maurov
@phibeck thank you for inviting me to review this submission to JOSS. Yes, I am happy to accept it.
