### 2023-09-22 #5868 @editorialbot
[REVIEW]: Fitspy: A Python package for spectral decomposition
**Submitting author:** @patquem (<a href="http://orcid.org/0009-0008-6936-1249">Patrick Quéméré</a>)
**Repository:** https://github.com/CEA-MetroCarac/fitspy
**Branch with paper.md** (empty if default branch):
**Version:** 2024.4
