### 2023-09-28 #5868 @patquem
Hello @phibeck,
Many thanks for your follow-up and reminders.
As a potential reviewer, I forgot to mention @teuben, who had few words about my draft the day I posted it (see https://github.com/openjournals/joss-reviews/issues/5655). He was interested in the Fitspy tool and I just saw he is listed as a reviewer. Feel free to invite him to review the paper.
