### 2023-09-29 #5868 @phibeck
Thanks for the suggestion @patquem. Since we're still at an early stage of the review, adding another reviewer probably won't delay the process significantly.

:wave: @teuben, would you be willing to review this submission for JOSS?
