### 2023-10-31 #5868 @maurov
@patquem @phibeck

I have started reviewing `fitspy` by reading the paper and the (short) documentation. I will install and play with the examples (plus trying some of my own data) soon, but I would like to express my general comments first.

A graphical interface on top of [lmfit](https://lmfit.github.io/lmfit-py/) may be of interest for those users who want to build a quick fitting model and apply it on multiple similar spectra, without prior knowledge of Python programming. I support this idea, but this software should be put in the right context and clearly stated in the paper/documentation. In fact, in addition to what here is called "spectral decomposition", "peak fitting" or "curve fitting" is performed in any field of science. This means that there already many (open source) softwares available and almost every researcher/community has a preferred tool. Clarifying the specific problems `fitspy` solves (a GUI for Lmfit?) and to which research community is targeted (X-ray fluorescence?) is mandatory. For example, why not comparing `fitspy` design to [Fityk](https://fityk.nieto.pl/)? Why two fancy astrophysics tools (`specutils` and `pyspeckit`) are cited (apart the obvious fact that they do peak fitting)?

In conclusion, the current "Statement of need" is too generic and unclear. I am asking myself if there is "substantial scholarly effort" behind this project. Is this a "yet another peak fitting tool"?
