### 2023-10-31 CEA-MetroCarac/fitspy#9 @maurov
documentation missing how to run the tests and examples
@patquem

I am evaluating `fitspy` (https://github.com/openjournals/joss-reviews/issues/5868) but I cannot run the tests and examples. The documentation should be updated with clear instructions and possibly a "quickstart tutorial". Please, could it be possible to push the documentation to a static web site? For example, [GitHub Pages](https://pages.github.com/) or [Read the docs](https://readthedocs.com/)

Tested on Ubuntu WSL ([Windows Subsystem for Linux](https://en.wikipedia.org/wiki/Windows_Subsystem_for_Linux)), python 3.9 in a fresh conda environment.

**Fitspy installation**

`pip install fitspy` -> OK

**Tests**

```bash
cd
git clone https://github.com/CEA-MetroCarac/fitspy.git
cd fitspy/tests
pip install pytest
```

All tests fail because the `examples` are not found. Error trace below:

```python
(fitspy) mauro@LAPCNRSMR:~/devel/fitspy/tests$ pytest
=============================================== test session starts ================================================
rootdir: /home/mauro/devel/fitspy
collected 0 items / 4 errors

====================================================== ERRORS ======================================================
____________________________________ ERROR collecting tests/test_gui_2d_maps.py ____________________________________
ImportError while importing test module '/home/mauro/devel/fitspy/tests/test_gui_2d_maps.py'.
Traceback:
../../../local/mambaforge/envs/fitspy/lib/python3.9/importlib/__init__.py:127: in import_module
test_gui_2d_maps.py:4: in <module>
    from examples.ex3_gui_2d_maps import gui_2d_maps
E   ModuleNotFoundError: No module named 'examples'
________________________ ERROR collecting tests/test_gui_apply_model_to_all_new_spectra.py _________________________
ImportError while importing test module '/home/mauro/devel/fitspy/tests/test_gui_apply_model_to_all_new_spectra.py'.
Traceback:
../../../local/mambaforge/envs/fitspy/lib/python3.9/importlib/__init__.py:127: in import_module
test_gui_apply_model_to_all_new_spectra.py:4: in <module>
    from examples.ex2_gui_apply_model_to_all_new_spectra import \
E   ModuleNotFoundError: No module named 'examples'
______________________________ ERROR collecting tests/test_gui_auto_decomposition.py _______________________________
ImportError while importing test module '/home/mauro/devel/fitspy/tests/test_gui_auto_decomposition.py'.
Traceback:
../../../local/mambaforge/envs/fitspy/lib/python3.9/importlib/__init__.py:127: in import_module
test_gui_auto_decomposition.py:4: in <module>
    from examples.ex0_gui_auto_decomposition import gui_auto_decomposition
E   ModuleNotFoundError: No module named 'examples'
_____________________________ ERROR collecting tests/test_inline_auto_decomposition.py _____________________________
ImportError while importing test module '/home/mauro/devel/fitspy/tests/test_inline_auto_decomposition.py'.
Traceback:
../../../local/mambaforge/envs/fitspy/lib/python3.9/importlib/__init__.py:127: in import_module
test_inline_auto_decomposition.py:5: in <module>
    from examples.ex4_inline_auto_decomposition import inline_auto_decomposition
E   ModuleNotFoundError: No module named 'examples'
============================================= short test summary info ==============================================
ERROR test_gui_2d_maps.py
ERROR test_gui_apply_model_to_all_new_spectra.py
ERROR test_gui_auto_decomposition.py
ERROR test_inline_auto_decomposition.py
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 4 errors during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
================================================ 4 errors in 0.13s =================================================
```

**Examples**

Example 0 -> fails with error:

```bash
(fitspy) mauro@LAPCNRSMR:~/devel/fitspy/examples$ python ex0_gui_auto_decomposition.py
/home/mauro/local/mambaforge/envs/fitspy/lib/python3.9/site-packages/examples/data/spectra_1 is not an existing directory
Traceback (most recent call last):
  File "/home/mauro/devel/fitspy/examples/ex0_gui_auto_decomposition.py", line 34, in <module>
    gui_auto_decomposition()
  File "/home/mauro/devel/fitspy/examples/ex0_gui_auto_decomposition.py", line 22, in gui_auto_decomposition
    appli.auto_eval_all()
  File "/home/mauro/local/mambaforge/envs/fitspy/lib/python3.9/site-packages/fitspy/app/callbacks.py", line 739, in auto_eval_all
    current_fname = self.current_spectrum.fname
AttributeError: 'NoneType' object has no attribute 'fname'
```

Example 1 -> only empty GUI appears:

```bash
(fitspy) mauro@LAPCNRSMR:~/devel/fitspy/examples$ python ex1_gui_reload_model.py
````

![image](https://github.com/CEA-MetroCarac/fitspy/assets/545294/47a4c7d0-33c3-490e-881e-e74f62939b4a)

Example 2 -> fails,  data not found

```
(fitspy) mauro@LAPCNRSMR:~/devel/fitspy/examples$ python ex2_gui_apply_model_to_all_new_spectra.py
/home/mauro/local/mambaforge/envs/fitspy/lib/python3.9/site-packages/examples/data/spectra_2 is not an existing directory
```

![image](https://github.com/CEA-MetroCarac/fitspy/assets/545294/5d27d5ce-bc6c-4706-80c9-bd0e16272f2e)

Example 3 -> fails, data not found

```bash
(fitspy) mauro@LAPCNRSMR:~/devel/fitspy/examples$ python ex3_gui_2d_maps.py
Traceback (most recent call last):
  File "/home/mauro/devel/fitspy/examples/ex3_gui_2d_maps.py", line 36, in <module>
    gui_2d_maps()
  File "/home/mauro/devel/fitspy/examples/ex3_gui_2d_maps.py", line 19, in gui_2d_maps
    appli.add_items(fnames=[str_map, unstr_map])
  File "/home/mauro/local/mambaforge/envs/fitspy/lib/python3.9/site-packages/fitspy/app/callbacks.py", line 816, in add_items
    spectrum.load_profile(fname)
  File "/home/mauro/local/mambaforge/envs/fitspy/lib/python3.9/site-packages/fitspy/spectra.py", line 209, in load_profile
    dfr = pd.read_csv(self.fname, sep=r'\s+|\t|,|;| ', engine='python',
  File "/home/mauro/local/mambaforge/envs/fitspy/lib/python3.9/site-packages/pandas/io/parsers/readers.py", line 948, in read_csv
    return _read(filepath_or_buffer, kwds)
  File "/home/mauro/local/mambaforge/envs/fitspy/lib/python3.9/site-packages/pandas/io/parsers/readers.py", line 611, in _read
    parser = TextFileReader(filepath_or_buffer, **kwds)
  File "/home/mauro/local/mambaforge/envs/fitspy/lib/python3.9/site-packages/pandas/io/parsers/readers.py", line 1448, in __init__
    self._engine = self._make_engine(f, self.engine)
  File "/home/mauro/local/mambaforge/envs/fitspy/lib/python3.9/site-packages/pandas/io/parsers/readers.py", line 1705, in _make_engine
    self.handles = get_handle(
  File "/home/mauro/local/mambaforge/envs/fitspy/lib/python3.9/site-packages/pandas/io/common.py", line 863, in get_handle
    handle = open(
FileNotFoundError: [Errno 2] No such file or directory: '/home/mauro/local/mambaforge/envs/fitspy/lib/python3.9/site-packages/examples/data/2D_maps/ordered_map.txt'
```

Example 4 -> fails

```sh
(fitspy) mauro@LAPCNRSMR:~/devel/fitspy/examples$ python ex4_inline_auto_decomposition.py
Traceback (most recent call last):
  File "/home/mauro/devel/fitspy/examples/ex4_inline_auto_decomposition.py", line 64, in <module>
    inline_auto_decomposition(show_plots=True)
  File "/home/mauro/devel/fitspy/examples/ex4_inline_auto_decomposition.py", line 32, in inline_auto_decomposition
    print(f"\n\nfilename: {fname}")
UnboundLocalError: local variable 'fname' referenced before assignment
```

Thanks for your help
