### 2023-11-03 22f8f92
revisit example and data paths
