+++ examples/ex3_gui_2d_maps.py
@@ -3,9 +3,11 @@ Example of 2D maps loading
 """
 import os
 import tkinter as tk
+from pathlib import Path
 
 from fitspy.app.gui import Appli
-from fitspy import DATA
+
+DATA = Path(__file__).parent / "data"
 
 
 def gui_2d_maps(dirname_res=None):
