+++ examples/ex2_gui_apply_model_to_all_new_spectra.py
@@ -3,9 +3,11 @@ Example of spectra fitting through the GUI and a model applied to all
 """
 import os
 import tkinter as tk
+from pathlib import Path
 
 from fitspy.app.gui import Appli
-from fitspy import DATA
+
+DATA = Path(__file__).parent / "data"
 
 
 def gui_apply_model_to_all(dirname_res=None):
