+++ conftest.py
@@ -1,0 +1,3 @@
+"""
+module description
+"""
