+++ fitspy/__init__.py
@@ -1,5 +1,1 @@
-import pathlib
 
-ROOT = pathlib.Path(__file__).parent.parent
-EXAMPLES = ROOT / 'examples'
-DATA = EXAMPLES / 'data'
