+++ pyproject.toml
@@ -1,3 +1,7 @@
 [build-system]
 requires = ['setuptools>=42']
 build-backend = 'setuptools.build_meta'
+
+[tool.pytest.ini_options]
+minversion = "6.0"
+testpaths = ["tests"]
