### 2023-11-03 8ea94f1
add tests and examples execution section
