### 2023-11-03 CEA-MetroCarac/fitspy#9 @patquem
Hi @maurov,
Thank you for your feedback.
Indeed, proceeding as you did, I reproduced the problems. I haven't encountered these problems before with my IDE (PyCharm) which automatically adds the contents roots and source roots to the PYTHONPATH.
After some modifications to allow executions in a standard terminal, the issues have been fixed for me.
Let me know if you still have problems running the tests.
_(I also added some text in the README.md about tests and examples execution.)_
