### 2023-11-08 CEA-MetroCarac/fitspy#9 @maurov
@patquem thanks for fixing this. The tests all pass now for me.

**Example 0** -> OK
**Example 1** -> fails

```
(fitspy) mauro@LAPROVEZZI:~/devel/fitspy/examples$ python ex1_gui_reload_model.py
Traceback (most recent call last):
  File "/home/mauro/devel/fitspy/examples/ex1_gui_reload_model.py", line 20, in <module>
    gui_reload_model()
  File "/home/mauro/devel/fitspy/examples/ex1_gui_reload_model.py", line 16, in gui_reload_model
    fitspy_launcher(fname_json=fname)
  File "/home/mauro/local/conda/envs/fitspy/lib/python3.9/site-packages/fitspy/app/gui.py", line 431, in fitspy_                                          launcher
    appli.reload(fname_json=fname_json)
  File "/home/mauro/local/conda/envs/fitspy/lib/python3.9/site-packages/fitspy/app/callbacks.py", line 274, in r                                          eload
    self.spectra = Spectra.load(fname_json)
  File "/home/mauro/local/conda/envs/fitspy/lib/python3.9/site-packages/fitspy/spectra.py", line 885, in load
    spectrum.preprocess()
  File "/home/mauro/local/conda/envs/fitspy/lib/python3.9/site-packages/fitspy/spectra.py", line 192, in preproc                                          ess
    self.load_profile(self.fname)
  File "/home/mauro/local/conda/envs/fitspy/lib/python3.9/site-packages/fitspy/spectra.py", line 209, in load_pr                                          ofile
    dfr = pd.read_csv(self.fname, sep=r'\s+|\t|,|;| ', engine='python',
  File "/home/mauro/local/conda/envs/fitspy/lib/python3.9/site-packages/pandas/io/parsers/readers.py", line 948,                                           in read_csv
    return _read(filepath_or_buffer, kwds)
  File "/home/mauro/local/conda/envs/fitspy/lib/python3.9/site-packages/pandas/io/parsers/readers.py", line 611,                                           in _read
    parser = TextFileReader(filepath_or_buffer, **kwds)
  File "/home/mauro/local/conda/envs/fitspy/lib/python3.9/site-packages/pandas/io/parsers/readers.py", line 1448                                          , in __init__
    self._engine = self._make_engine(f, self.engine)
  File "/home/mauro/local/conda/envs/fitspy/lib/python3.9/site-packages/pandas/io/parsers/readers.py", line 1705                                          , in _make_engine
    self.handles = get_handle(
  File "/home/mauro/local/conda/envs/fitspy/lib/python3.9/site-packages/pandas/io/common.py", line 863, in get_h                                          andle
    handle = open(
FileNotFoundError: [Errno 2] No such file or directory: '..\\examples\\data\\spectra_2\\InP-1_42-P21.txt'
```

**Example 3** -> OK
**Example 4** -> OK

## Some further questions

1) I cannot horizontally extend the settings panel ((see screenshot)

![image](https://github.com/CEA-MetroCarac/fitspy/assets/545294/46174ec7-16dc-44bd-9852-64f33160860d)

2) When I use the top right cross to close the GUI window, the command line shell is stuck and CTRL+C does not work to close the process. I have to kill the whole command line shell or the python process itself. This is very annoying. Could you add a "Quit" button (possibly binded to CTRL+Q) to quit the GUI in a clean way?

3) Is it possible to show the fit residual (in a panel with scaled y scale)? **This is very important**.

4) The "Parameters" panel is very slow to show/refresh (any idea why?), it does not include the background model information and does not nicely render (see screenshot)

![image](https://github.com/CEA-MetroCarac/fitspy/assets/545294/d540484c-1c37-4332-8b07-ab9b1ca9ddc0)
