### 2023-11-08 c01ca0a
revisit 'sbar' adaptative width
