### 2023-11-08 98d8ef9
quit 'properly' the application
