### 2023-11-08 CEA-MetroCarac/fitspy#9 @patquem
Hi @maurov,

Thanks again for your feedback and sorry for the problems you encountered.

- Example 1 that failed : fixed (incompatibility between the Windows and the Linux separators of the input pathname in the model.json)
- Panel width: strange behaviour certainly due to your screen resolution (in my case, all are correctly aligned). I changed the fixed width with a self-adaptative one. Tell me if this solution allows you to have the entire panel visible.
- Exit the application properly: fixed (by clicking on the 'x' button at the top right).
- Fit residual displaying: this option is already available in the 'Figure settings' (deactivated by default). Note that the residuals are not saved in the models -> When you reload a model, you need to 'Fit' the model to visualize the residual.
-  Parameters panel - refreshing : for me it is not so slow (~0.1s - 0.2s), the time for the modified object to be correctly reinitialized and for the panel to be fully rewritten (considering that one or several models can be removed).
-  Parameters panel - background parameters missing : you are right. The background fitting has been added lately. (-> To add in a future release).
- Parameters panel - not nicely render: what's wrong in the screenshot ? This is how I did to make the parameters, their bounds and the associated expressions (constraints) visible in a single Tab .
The Application was created with Tkinter (not really the best choice for creating such a GUI).
Using PySide or PyQt would certainly improve the final rendering of the application. To address in a future release :).
Thank you again and don't hesitate to provide further feedback.
Patrick
