### 2023-11-09 95ba27d
bugfix: 'vary' not in model.param_hints
