### 2023-11-09 2184f21
manage bkg_model line and parameters displaying
