### 2023-11-10 12550b7
add get_bkg_model_name()
