### 2023-11-10 c85b453
remove Gaussian from BKG models
