### 2023-11-10 d1c4b94
define bkg_mode hint parameters from guess
