### 2023-11-10 CEA-MetroCarac/fitspy#9 @patquem
- Parameters panel - background parameters missing : **DONE** 959a68addf0db7bcb7dc57ff7ee4dd37fac55d39
