### 2023-11-10 6820173
show only used model parameters
