### 2023-11-10 CEA-MetroCarac/fitspy#9 @patquem
- Parameters panel - background parameters missing : **DONE** 959a68addf0db7bcb7dc57ff7ee4dd37fac55d39
- Parameters panel - not nicely render : **REVISITED** 682017350993e1fa962de89a881dd5c9bfbd1f22
@maurov if your remark was related to the empty parameter columns, I made changes to remove them. see the screenshot below, when 'fwhm (left)' and 'fwhm (right)' parameters are not used.

![image](https://github.com/CEA-MetroCarac/fitspy/assets/16154687/7315424c-8b5e-46b1-a653-64f56ebb4d58)
