### 2023-11-13 #5868 @patquem
Hi,
Thank you @maurov for your reviewing and your comments.
You mentioned a lack of clarity regarding the objectives of the software and I will try to clarify them.
Which communauty is this tool aimed at ? I would tell anyone who needs to do peak fitting (and only peak fitting) in the easiest and the quickest (in term of accessibility) way possible.
While I agree with you (there are already many other softwares that do peak fitting and Fitspy may appear like "yet another peak fitting tool"), most of the softwares with a GUI don't manage 2D-maps and/or do not have multiprocessing capabilities (see the (non exhaustive) list of open-source softwares below).
Although Fitik that you mentioned is a tool with an advanced GUI and is undeniably more evoluated that Fitspy, it does not have 2D-map and parallel capabilities.
The only ones that I have identified and that combine these 2 features are XISMUS and HyperspyUI.
From what I have seen, XISMUS is strongly attached to X-Ray Fluorescence datasets.
On this point, Fitspy aims to be a generic and agnostic tool.
HyperspyUI seems to be the closest tool but (in my experience) is not as easy to use as Fitspy and this tool aims at the end to adress all Hyperspy options, far from the only need of peak fitting to cover.
After testing Fitspy, @maurov do you agree to consider Fitspy as satisfying some of the terms used in the "scholarly effort" section :  faster and/or easier and/or simpler than most open source software already existing?
If so, I will rewrite the "Statement of need", with emphasis on consideration about the dedicated easy-to-use peak fitting tool and its 2D-map/parallel capabilities.
![image](https://github.com/openjournals/joss-reviews/assets/16154687/5083cbcc-9b31-4e23-9e9a-ac3e8c44a8bf)
