+++ README.md
@@ -25,7 +25,7 @@ The peak models considered in fitspy are :
 * `Asymetric Lorentzian`
 * `Pseudovoigt`
 
-A `constant`, `linear`, `parabolic`, `gaussian` or `exponential` background can
+A `constant`, `linear`, `parabolic` or `exponential` background can
 also be added in the fitting.
 
 All actions allowed with the GUI can be easily executed in script mode (see
@@ -34,7 +34,6 @@ These actions (like baseline definition and removal, peaks definition,
 parameters constraints, ...) can be saved in a 'fitspy' `model` and replayed as
 is or applied to other new spectra.
 
-See the [documentation](https://github.com/CEA-MetroCarac/fitspy/tree/main/doc) for more details.
 
 ### Installation
 
@@ -42,23 +41,38 @@ See the [documentation](https://github.com/CEA-MetroCarac/fitspy/tree/main/doc) 
 pip install fitspy
 ```
 
-### Examples and tests execution
+### Tests and examples execution
 
-First you need to clone the project:
-```
-git clone https://github.com/CEA-MetroCarac/fitspy.git
-```
-To launch an example (here *ex0_gui_auto_decomposition.py*):
-```
-cd fitspy
-python example/ex0_gui_auto_decomposition.py
-```
-To execute the tests:
 ```
 pip install pytest
+git clone https://github.com/CEA-MetroCarac/fitspy.git
 cd fitspy
 pytest
+python example/ex0_gui_auto_decomposition.py
+...
 ```
+
+### Quick start
+
+Launch the application:
+```
+fitspy
+```
+Then, from the top to the bottom of the right panel:
+
+- `Select` file(s)
+- <span style="color: rgba(0, 0, 0, 0.3);">*(Optional)* Define the **X-range**</span>
+- Define the baseline to `subtract` *(On the figure: left-click to add a baseline point / right-click to delete a baseline point)*
+- <span style="color: rgba(0, 0, 0, 0.3);">*(Optional)* Normalize the spectrum/spectra</span>
+- Click on the `Peaks` panel to activate it
+- Select `Peak model` and add peaks *(On the figure: left-click to add a peak / right-click to delete a peak)*
+- <span style="color: rgba(0, 0, 0, 0.3);">*(Optional)* Add a background (**BKG model**) to be fitted</span>
+- <span style="color: rgba(0, 0, 0, 0.3);">*(Optional)* Use **Paramaters** to set bounds and constraints</span>
+- `Fit` the selected spectrum/spectra *(fitted parameters are automatically saved in the related *'filename'.csv* )*
+- <span style="color: rgba(0, 0, 0, 0.3);">*(Optional)* **Save** the model to be replayed later</span>
+
+See the [documentation](https://github.com/CEA-MetroCarac/fitspy/tree/main/doc) for more details.
+
 
 ### Authors informations
 
