### 2023-11-24 #5868 @maurov
@patquem I apologize for not being very reactive on this review, but I am very busy these days. I hope to move forward in the next two weeks.

As said in my previous comment, you should clearly give in the "Statement of need" which scientific community this code is targeting and give a real application example. I understand that you want to provide a generic and efficient peak fitting tool, but it is impossible to get all features for all users communities, so the point here is not to demonstrate which software is the best, but just to provide a given tool to a given community. For example, when you refer to 2D maps, are you referring to X-ray fluorescence maps? For such application, one very good open source software that is not in your table is [PyMca](http://www.silx.org/doc/PyMca/dev/index.html).
