### 2023-12-18 #5868 @phibeck
Thanks @maurov for getting the review started! Have your issues in https://github.com/CEA-MetroCarac/fitspy/issues/9 been addressed? It sounds like there are some clear steps for @patquem to improve the manuscript, too.

@FCMeng could you provide an update on the progress of your review? Thank you!
