### 2023-12-18 #5868 @patquem
Hello @maurov,
Thank you again for your comments.
As previously said, Fitspy (like Fityk) is a generic and agnostic tool dedicated to do peak fitting and only peak fitting and is not intended for any specific community or specific applications.
To justify my point and answer to your questions, I would say I work in a multidisciplinary institute working on many fields like biology, batteries and semi-conductor with a wide range of characterization techniques. Over the last 5-6 months, Fitspy has been used to process spectra issued from Raman, PL, XPS, XRD and EELS to characterize materials from the range of several microns to the range of thousands of mm (wafers). The number of spectra processed by Fitspy during these studies varied from a few to several hundred thousand. The 2D map characterisation was used both for the characterisation of PL spectra (at the wafer scale) and to process EELS spectra at the micrometer scale.
Fitspy has been used by both experienced and interns who have appreciated the tool for its “easy-to-use” compared to the reference tools they currently use (as they said: “reference tools offer a large range of services but, as a consequence, are more complex and not so ergonomic to use for such basic and (often) repetitive tasks that peak fitting operations are”).
**These are the reasons why it does not seem appropriate to me to focus the Fitspy “statement of need” on a particular community and specific applications.**
