### 2023-12-18 #5868 @patquem
@maurov @FCMeng and @phibeck :
**With the revisited Statement of need below: Does the article have a chance of being published or do you consider it as definitively rejected ?**
Indeed, I need an answer because articles issued from applicative studies with Fitspy are now about to be submitted and authors need to know if a Fistpy reference does exist (Initially, in July, when I submitted the present article to JOSS, I was thinking IT could be the reference paper).
Thank you.
