### 2023-12-18 #5868 @patquem
**Statement of need (revisited)**

Numerous open-source tools for spectral fitting exist. However, most of them have been designed for specific application domains, offering a broad range of services beyond mere spectral fitting. Consequently, these tools can prove challenging to use, especially for less experienced individuals.

In the vein of generic tools like **Fityk** or **PRISMA**, Fitspy is a dedicated tool for spectral fitting—and only spectral fitting—with the following characteristics or functionalities:

**Agnostic Nature**: Fitspy is not tied to any specific physical quantity or database. It is designed to process spectra regardless of their x-support and y-intensity without requiring prior knowledge.

**Python Implementation**: Fitspy is coded in Python. As a result, spectra can be easily processed using Python scripts, catering to individuals with basic knowledge of the language.

**2D Maps**: Fitspy has been designed to handle spectra derived from 2D acquisitions. Note that beyond "2D", dimensions can encompass time or any other dimension. When dealing with 2D data, an interactive map in the Fitspy GUI allows users to locate and select spectra of interest easily.

**Multiprocessing Capabilities**: Fitspy enables spectral fit processing on multiple processors, enhancing efficiency.

**Constrained Parameters**: Leveraging the **Lmfit** library, Fitspy empowers users to impose constraints on parameter ranges or establish constraints between parameters using literal expressions.

**Simple GUI**: Fitspy has been designed to be as intuitive and simple to use as possible (subjective criterion).

To the author's knowledge, although many open-source software applications are more advanced in certain aspects mentioned, none of them appears to encompass all the functionalities described above.
