### 2023-12-19 CEA-MetroCarac/fitspy#9 @maurov
@patquem thanks for implementing most of my suggestions regarding this issue while running the examples.
