### 2023-12-19 CEA-MetroCarac/fitspy#10 @maurov
documentation and examples improvements
As follow-up to [my comments in JOSS review](https://github.com/openjournals/joss-reviews/issues/5868) I propose to improve the documentation and examples by taking into account the following points (no particular order/priority and the list is not exhaustive nor complete):

- [x] Baseline: how to fit the baseline only?
- [x] Baseline/GUI: the "Auto" button adds always a linear baseline, even if the polynomial is selected
- [x] What is the difference between Baseline and "BKG model" in the Peaks section? To me, the baseline is the background model and one should be able to fit it before start adding the peak(s) models to the fit.
- [ ] GUI: would be nice to have the possibility to select multiple X-range regions to exclude/include temporary in the fit.
- [x] The baseline should be clearly visible in the fits. For example, the left plot of `ex4_inline_auto_decomposition.py`  should show the baseline.
- [x] The ASCII input format of the 2D data should be clearly specified. Just pointing to the raw data is not enough and very annoying to dig into it.
- [x] In `ex3_gui_2d_maps.py` when clicking on the images the data are shown in the main panel, but not the fits.
- [x] Why normalization to 100? Usually one normalize to 1.

Please, no need to reply point-by-point here, as these are ideas only (not mandatory improvements). If you want to take them into account, I propose to commit your changes to the code/documentation/examples directly into the repository master branch.
