### 2023-12-19 #5868 @maurov
@patquem

thanks for the revised statement of need and I also appreciate the paragraph showing a use case for Fitspy. This is what I was proposing since the beginning: say how your current users use this software for. Please, insert them directly into the revised manuscript and regenerate it.

@patquem @phibeck

I acknowledge the effort for providing this peak-fitting tool to the community and I think this work fits the scope and audience of JOSS. I may be in favor of accepting this contribution, but only once the documentation is improved and extended. I have opened a [specif issue for it](https://github.com/CEA-MetroCarac/fitspy/issues/10) with some ideas, even though my time is very limited and the list is far from being complete and exhaustive. Overall, the provided "README" file mainly describing what the buttons on the GUI do is not enough.

@phibeck

I went trough all the points, but some of them are still not checked because are currently not met, in my opinion.
