### 2023-12-19 #5868 @phibeck
Thanks @patquem and @maurov for the updates. Regarding the scholarly effort (and after consulting with other editors), it is not so much a question of whether there are alternative packages, but whether the amount of work done by the authors meets the minimal requirements. The editors already reviewed the scope in a first round and found that it meets the requirements, but if reviewers have strong doubts these should be stated.

Given the assessment by @maurov it sounds like the article can be published once the issues are taken care of by @patquem. However, we need at least a second report, and I hope @FCMeng is still available to give one.
