+++ fitspy/spectrum.py
@@ -1,0 +1,533 @@
+"""
+Classes dedicated to spectra fitting
+"""
+import os
+# import warnings
+import itertools
+# import csv
+from copy import deepcopy
+from concurrent.futures import ProcessPoolExecutor
+import numpy as np
+import pandas as pd
+# import matplotlib.pyplot as plt
+# from matplotlib.widgets import RangeSlider
+from scipy.signal import find_peaks
+# from scipy.interpolate import interp1d
+# from scipy.ndimage import gaussian_filter1d
+from lmfit import Model, report_fit, Parameters
+from lmfit.model import ModelResult
+from lmfit.models import ConstantModel, LinearModel, ParabolicModel, \
+    ExponentialModel  # pylint:disable=unused-import
+
+from fitspy.utils import closest_index
+from fitspy.utils import save_to_json, load_from_json
+# from fitspy.models import gaussian, lorentzian, gaussian_asym, lorentzian_asym
+# from fitspy.models import pseudovoigt
+from fitspy.app.utils import convert_dict_from_tk_variables
+from fitspy.app.utils import dict_has_tk_variable
+from fitspy.baseline import BaseLine
+from fitspy import MODELS, BKG_MODELS
+
+
+# MODELS = {"Gaussian": gaussian,
+#           "Lorentzian": lorentzian,
+#           "PseudoVoigt": pseudovoigt,
+#           "GaussianAsym": gaussian_asym,
+#           "LorentzianAsym": lorentzian_asym}
+#
+# BKG_MODELS = ['None', 'Constant', 'Linear', 'Parabolic', 'Exponential']
+#
+# KEYS = ['x0', 'ampli', 'fwhm', 'fwhm_l', 'fwhm_r', 'alpha']
+
+
+class Spectrum:
+    """
+    Class dedicated to spectrum processing
+
+    Attributes
+    ----------
+    fname: str
+        Filename associated to spectrum to handle
+    range_min, range_max: floats
+        Range associated to the spectrum support to work with
+    norm_mode: str
+        Mode used for the spectrum normalization ('Maximum' or 'Attractor')
+    norm_position_ref: float
+        Reference position associated to the 'Attractor' mode during the
+        normalization
+    x0, y0: numpy.array((n))
+        Arrays related to spectrum raw support and intensity (resp.)
+    x, y: numpy.array((n))
+        Arrays related to spectrum modified support and intensity (resp.)
+    peaks: list of ints
+        Indices of peaks
+    peaks_params: dict
+        Dictionary passed to scipy.signal.find_peaks for peaks determination.
+        Could be a dictionary with Tkinter.Variable as values.
+    models: list of lmfit.Model
+        List of the lmfit models
+    models_labels: list of str
+        List of labels associated to the models. Default is ['1', '2', '3', ...]
+    models_index: itertools.count
+        Counter used for models indexing when creating a lmfit.Model
+    bkg_model: lmfit.Model
+        Background model to fit with the composite peaks models, among :
+        [None, 'ConstantModel', 'LinearModel', 'ParabolicModel',
+        'ExponentialModel']
+    fit_method: str
+        Method used for fitting. See lmfit.Model.fit().
+        Default method is 'leastsq'.
+    fit_negative: bool
+        Activation keyword to take into account negative values when fitting.
+        Default is False.
+    max_ite: int
+        Maximum number of iteration associated to the fitting process.
+        An iteration consists in evaluating all the 'free' parameters once.
+        Default is 200.
+    result_fit: lmfit.ModelResult
+        Object resulting from lmfit fitting
+    baseline: Baseline object
+        Baseline associated to the spectrum (to be removed)
+    baseline_history: list of list
[…]
