### 2024-01-03 0ff610a
bugfix: index not a str in models_labels
