### 2024-01-03 adaafb1
display models params in the 2D-map
