+++ fitspy/spectra_map.py
@@ -5,10 +5,14 @@ import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
 from matplotlib.widgets import RangeSlider
+from parse import Parser
 
 from fitspy.utils import closest_index
 from fitspy.spectra import Spectra
 from fitspy.spectrum import Spectrum
+
+POLICY = "{name}  X={x}  Y={y}"
+PARSER = Parser(POLICY)
 
 
 class SpectraMap(Spectra):
@@ -28,19 +32,23 @@ class SpectraMap(Spectra):
         (xmin, xmax, ymin, ymax)
     coords: list of list of 2 floats
         List containing the (x,y) coordinates for each spectrum in the 2D-map
-    arr: numpy.ndarray(((len(shape_map[0]) * len(shape_map[1]), n))
+    intensity: numpy.ndarray(((len(shape_map[0]) * len(shape_map[1]), n))
         Array of spectra intensities (n-values) associated to the xy_map coords
+    arr: numpy.ndarray((len(shape_map[0]), len(shape_map[1]))
+        Array associated to the xy_map coords
     ax: Matplotlib.pyplot.axis
         Axis associated to the 2D-map displaying
     img: Matplotlib.image.AxesImage
         Image related to the 2D-map displaying
     cbar: Matplotlib.pyplot.colorbar
         Colorbar associated to the spectra 2D-map displaying
+    ax_slider: Matplotlib.pyplot.axis
+        Axis associated to the RangeSlider object
+    slider: Matplotlib.widgets.RangeSlider
+        Slider associated to the intensity integration range or to a model
+        parameter in the 2D-map displaying
     xrange: tuple of 2 floats
         Range of intensity to consider in the 2D-map displaying
-    slider: Matplotlib.widgets.RangeSlider
-        Slider associated to the intensity integration range in the 2D-map
-        displaying
     """
 
     def __init__(self):
@@ -50,11 +58,13 @@ class SpectraMap(Spectra):
         self.shape_map = None
         self.extent = None
         self.coords = None
+        self.intensity = None
         self.arr = None
 
         self.ax = None
         self.img = None
         self.cbar = None
+        self.ax_slider = None
         self.slider = None
         self.xrange = None
 
@@ -84,7 +94,7 @@ class SpectraMap(Spectra):
         for vals in arr[1:]:
             # create the related spectrum object
             spectrum = Spectrum()
-            spectrum.fname = f"{fname}  X={vals[1]}  Y={vals[0]}"
+            spectrum.fname = POLICY.format(name=fname, x=vals[1], y=vals[0])
             intensity = vals[2:][inds]
             spectrum.x = x
             spectrum.y = intensity
@@ -98,8 +108,18 @@ class SpectraMap(Spectra):
         self.xy_map = (x_map, y_map)
         self.shape_map = (len(self.xy_map[1]), len(self.xy_map[0]))
         self.extent = extent
-        self.arr = intensity_map
+        self.intensity = intensity_map
+        self.arr = np.sum(intensity_map, axis=1).reshape(self.shape_map)
         self.coords = coords
+
+    def spectrum_indices(self, spectrum):
+        """ Return the (i, j) map indices related to 'spectrum' """
+        res = PARSER.parse(spectrum.fname)
+        x = float(res.named['x'])
+        y = float(res.named['y'])
+        j = self.xy_map[0].index(x)
+        i = self.xy_map[1].index(y)
+        return i, j
 
     def plot_map(self, ax):
         """ Plot the integrated spectra map intensities on 'ax' """
@@ -108,34 +128,47 @@ class SpectraMap(Spectra):
 
         fig = self.ax.get_figure()
         fig.subplots_adjust(top=0.92)
[…]
