+++ setup.py
@@ -13,6 +13,7 @@ setup(
         "matplotlib",
         "scipy",
         "lmfit",
+        "parse",
         "pywin32; platform_system == 'Windows'",
     ],
     packages=find_packages(where='.', include=['fitspy*']),
