+++ doc/README.md
@@ -6,35 +6,26 @@ The main class of Fitspy is [`Spectra`](../fitspy/spectra.py) (inherited
 from `list`) that contains a list of  [`Spectrum`](../fitspy/spectra.py)
 objects.
 
-For each of these [`Spectrum`](../fitspy/spectra.py) objects, the corresponding (x, y) profile is associated.<br>
-2D maps of spectra (named `SpectraMap`) can also be given as input according to a dedicated format. They consist in a set of `Spectrum` with associated grid coordinates (see below).
+For each of these [`Spectrum`](../fitspy/spectra.py) objects, the corresponding (x, y) profile is associated.
+
+Additionally, 2D maps of spectra (named `SpectraMap`) can be provided as input according to a dedicated format. They consist in a set of `Spectrum` with associated grid coordinates (see below).
 
 ![](_static/spectra.png)
 
-In term of processing, given the users instructions, `baselines`
-can be defined, as well as a list of `peaks` models (Gaussian, Lorentzian
-, ...) used to fit the spectra profiles.
+The spectra processing are based on  `baselines` definition (optional) and `peaks` models definition (Gaussian, Lorentzian, ...) used to fit the spectra profiles.
 
-In term of usage, to avoid confusing situations,
-it is recommanded to address spectra processing from the top to the bottom frames as depicted hereafter.
-
+**Important: In terms of usage, to avoid confusion, it is recommended to address spectra processing from the top to the bottom frames as depicted hereafter**.
 
 ## Preliminary remarks
 
-**Axis limits**<br> 
-To ease the comparison between spectra, the (x, y) axis limits are
-defined from the first spectrum displayed on the figure.<br>
-Then, they are applied to all the following ones.<br>
-To readapt the axis limits to the current spectrum click
-on <img src="_static/home.png" width="20"> of the navigation toolbar
-(auto-rescaling).
+**Axis limits**<br>
+To facilitate the comparison between spectra visually, the (x, y) axis limits in the figure are set based on the first displayed spectrum.<br>
+Subsequently, these limits are applied to all the following spectra.<br>
+To readjust the axis limits to the current spectrum click on <img src="_static/home.png" width="20"> of the navigation toolbar (auto-rescaling).
 
-**Mouse management**<br> 
-click buttons with the mouse is used with the
-navigation toolbar for zooming and translating as well as for defining
-baseline points or peaks points.<br>
-The activation/deactivation of the mouse associated to these 2 last modes (baseline and peaks selection) requires the corresponding
-frames to be "clicked" beforehand.
+**Mouse management**<br>
+Mouse click buttons are used both with the navigation toolbar (for zooming and translating) and for defining baseline points or peaks points.<br>
+The activation/deactivation of the mouse associated with these last 2 modes (baseline and peaks selection) requires the corresponding frames to be "clicked" beforehand.
 
 ## Files selection
 
@@ -44,27 +35,18 @@ The files selection is realized thanks to the following widget:
 
 Two types of data can be interpreted.
 
-* The first one consists in input data files that contains 2 columns
-  related to respectively the support (x) and the spectrum intensity (y).
-  The separators between the columns can be the tabulation `\t`, the coma
-  `,`, the semicolon `;` or the space ` `. Note that the first row is skipped
-  and the (x, y) rows can be unordered.
+* The first type consists of input data files containing 2 columns representing the support (x) and spectrum intensity (y) respectively.
+  The separators between the columns can be tabulation `\t`, comma `,`, semicolon `;` or space ` `. Note that the first row is skipped and the (x, y) rows can be unordered.
 
 
-* The second one is related to **2D-map** input data. the first line of the
-  file is related to the support (x) of the spectra and starts with 2
[…]
