### 2024-01-04 b7cfe36
bugfix: bad import in docstring 'plot'
