+++ paper/paper.bib
@@ -1,3 +1,33 @@
+@software{Fityk,
+        author = "Wojdyr, Marcin",
+        title = "{{\it Fityk}: a general-purpose peak fitting program}",
+        journal = "Journal of Applied Crystallography",
+        year = "2010",
+        volume = "43",
+        number = "5 Part 1",
+        pages = "1126--1128",
+        month = "Oct",
+        doi = {10.1107/S0021889810030499},
+        url = {https://doi.org/10.1107/S0021889810030499},
+        abstract = {{\it Fityk} is portable, open-source software for nonlinear curve fitting and data analysis. It specializes in fitting a sum of bell-shaped functions to experimental data. In particular, it enables Pawley refinement of powder diffraction data and size{--}strain analysis.},
+        keywords = {curve fitting, Pawley refinement},
+}
+
+@software{PRISMA,
+        author = {Flores, Eibar and Mozhzhukhina, Nataliia and Li, Xinyu and Norby, Poul and Matic, Aleksandar and Vegge, Tejs},
+        title = {PRISMA: A Robust and Intuitive Tool for High-Throughput Processing of Chemical Spectra**},
+        journal = {Chemistry–Methods},
+        volume = {2},
+        number = {10},
+        pages = {e202100094},
+        keywords = {electrochemical energy storage, high throughput analysis, in situ spectroscopy, operando spectroscopy, software},
+        doi = {https://doi.org/10.1002/cmtd.202100094},
+        url = {https://chemistry-europe.onlinelibrary.wiley.com/doi/abs/10.1002/cmtd.202100094},
+        eprint = {https://chemistry-europe.onlinelibrary.wiley.com/doi/pdf/10.1002/cmtd.202100094},
+        abstract = {Abstract Here we introduce PRISMA, an open-source tool to rapidly analyse large numbers of spectra into meaningful spectroscopic trends. PRISMA follows a human-in-the-loop workflow, where the user interacts with an intuitive graphical user interface to control multiple steps in the spectrum analysis process: trimming, baseline correction, and peak fitting. The tool outputs the results in an easy-to-read csv format within seconds. We describe the functionalities implemented in PRISMA and test its capabilities with three experimental cases relevant to the study of Electrochemical energy storage and conversion devices: a temperature-dependent Raman measurement of phase transitions, a linear Raman mapping of a graphite composite electrode, and an operando X-ray diffraction experiment of lithium nickel oxide composite battery electrode. The case studies demonstrate the robustness of the app and its ability to unearth insightful trends in peak parameters, which are easier to represent, interpret and further analyse with more advanced techniques.},
+        year = {2022}
+}
+
 @software{lmfit,
   author       = {Newville, Matthew and
                   Stensitzki, Till and
@@ -52,32 +82,18 @@
   url          = {https://doi.org/10.5281/zenodo.8092679}
 }
 
-@software{specutils,
-       author = {{Astropy-Specutils Development Team}},
-        title = "{Specutils: Spectroscopic analysis and reduction}",
-     keywords = {Software},
- howpublished = {Astrophysics Source Code Library, record ascl:1902.012},
-         year = 2019,
-        month = feb,
-          eid = {ascl:1902.012},
-        pages = {ascl:1902.012},
-archivePrefix = {ascl},
-       eprint = {1902.012},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2019ascl.soft02012A},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@software{pyspeckit,
-       author = {{Ginsburg}, Adam and {Mirocha}, Jordan},
[…]
