+++ paper/paper.md
@@ -22,11 +22,11 @@ bibliography: paper.bib
 
 **`Fitspy`** is a dedicated tool to **fit** **sp**ectra in **py**thon.
 
-Spectrum decomposition (also known as fitting) is the choice of a linear combination of peak models to best represent an experimental spectrum. `fitspy` currently has implementations for: Gaussian, Lorientzan, related asymmetric functions and the Pseudovoigt model. Spectrum decomposition corresponds to the resolution of a minimization problem, which can be subject to bounds and constraints on model parameters.
+Spectrum decomposition (also known as fitting) is the choice of a linear combination of peak models to best represent an experimental spectrum. `Fitspy` currently has implementations for: Gaussian, Lorientzan, related asymmetric functions and the Pseudovoigt model. Spectrum decomposition corresponds to the resolution of a minimization problem, which can be subject to bounds and constraints on model parameters.
 
-The `fitspy` GUI (see \autoref{fig:GUI}) is designed to be as simple and intuitive to use as possible.
+The `Fitspy` GUI (see \autoref{fig:GUI}) is designed to be as simple and intuitive to use as possible.
 
-![The `fitpsy` GUI.\label{fig:GUI}](https://cea-metrocarac.github.io/fitspy/fitspy.png){width=85%}
+![The `Fitpsy` GUI.\label{fig:GUI}](https://cea-metrocarac.github.io/fitspy/fitspy.png){width=85%}
 
 # Statement of need
 
@@ -34,15 +34,18 @@ The analysis of spectra in many areas of physics, from materials characterisatio
 into more or less complex models to estimate the chemical composition of the subject being analysed.
 To carry out these decompositions, research communities can rely on spectral decomposition tools provided in numerous commercial or open source softwares.
 
-Although commercial softwares are often of very high quality and high performance, they are by definition subject to licensing
-which is likely to limit their use within a team of researchers.
-In addition, the large amount of functionalities often proposed in these softwares can sometimes make them difficult to use.
-Finally, commercial softwares do not always make it easy to export the results in a open format (like .csv), or to save all the processing steps to be modified at a later time.
-These 2 last aspects, among others, severely limit the exchange of results and models within research teams as well as the use and re-use of models to compare analogous sets of spectra.
+Numerous open-source tools for spectral fitting exist. However, most of them have been designed for specific application domains, offering a broad range of services beyond mere spectral fitting. Consequently, these tools can prove challenging to use, especially for less experienced individuals.
 
-On the other hand, in the world of Open Source and particularly in the world of Python programming, libraries such as `lmfit` (cf. @lmfit), `specutils` (cf. @specutils) or `pyspeckit` (cf. @pyspeckit) offer everything that is needed to carry out spectral decomposition. However, their implementations require knowledge in programming, which limits their use by a large groups of people.
+In the vein of generic tools like `Fityk` or `PRISMA` (cf. @Fytik @PRISMA), `Fitspy` is a dedicated tool for spectral fitting — and only spectral fitting — with the following characteristics or functionalities:
+* **Agnostic Nature**: `Fitspy` is not tied to any specific physical quantity or database. It is designed to process spectra regardless of their x-support and y-intensity without requiring prior knowledge.
+* **Python Implementation**: `Fitspy` is coded in Python. As a result, spectra can be easily processed using Python scripts, catering to individuals with basic knowledge of the language.
+* **2D Maps**: `Fitspy` has been designed to handle spectra derived from 2D acquisitions. Note that the beyond "2D", dimensions can encompass time or any other dimension. When dealing with 2D data, an interactive map in the `Fitspy` GUI allows users to locate and select spectra of interest easily.
+* **Multiprocessing Capabilities**: `Fitspy` enables spectral fit processing on multiple processors, enhancing efficiency.
[…]
