+++ paper/paper.md
@@ -32,19 +32,24 @@ The `Fitspy` GUI (see \autoref{fig:GUI}) is designed to be as simple and intuiti
 
 The analysis of spectra in many areas of physics, from materials characterisation to astrophysics, often requires their decomposition
 into more or less complex models to estimate the chemical composition of the subject being analysed.
-To carry out these decompositions, research communities can rely on spectral decomposition tools provided in numerous commercial or open source softwares.
+To carry out these decompositions, research communities can rely on spectral decomposition tools.
+Whilr many open-source tools for spectral fitting exist, most of them have, however, been designed for specific application domains, offering a broad range of services beyond mere spectral fitting. Consequently, these tools can prove challenging to use, especially for less experienced individuals.
 
-Numerous open-source tools for spectral fitting exist. However, most of them have been designed for specific application domains, offering a broad range of services beyond mere spectral fitting. Consequently, these tools can prove challenging to use, especially for less experienced individuals.
+In the vein of generic tools like `Fityk` [@Fityk] or `PRISMA`  [@PRISMA], `Fitspy` is a dedicated tool for spectral fitting — and only spectral fitting — with the following characteristics or functionalities:
 
-In the vein of generic tools like `Fityk` or `PRISMA` (cf. @Fytik @PRISMA), `Fitspy` is a dedicated tool for spectral fitting — and only spectral fitting — with the following characteristics or functionalities:
 * **Agnostic Nature**: `Fitspy` is not tied to any specific physical quantity or database. It is designed to process spectra regardless of their x-support and y-intensity without requiring prior knowledge.
+
 * **Python Implementation**: `Fitspy` is coded in Python. As a result, spectra can be easily processed using Python scripts, catering to individuals with basic knowledge of the language.
+
 * **2D Maps**: `Fitspy` has been designed to handle spectra derived from 2D acquisitions. Note that the beyond "2D", dimensions can encompass time or any other dimension. When dealing with 2D data, an interactive map in the `Fitspy` GUI allows users to locate and select spectra of interest easily.
+
 * **Multiprocessing Capabilities**: `Fitspy` enables spectral fit processing on multiple processors, enhancing efficiency.
-* **Constrained Parameters**: Leveraging the `lmfit` library (cf. @lmfit), `Fitspy` empowers users to impose constraints on parameter ranges or establish constraints between parameters using literal expressions.
+
+* **Constrained Parameters**: Leveraging the `lmfit` library [(@lmfit], `Fitspy` empowers users to impose constraints on parameter ranges or establish constraints between parameters using literal expressions.
+
 * **Simple GUI**: `Fitspy` has been designed to be as intuitive and simple to use as possible (subjective criterion).
 
-To the author's knowledge, although many open-source software applications are much more advanced in certain aspects mentioned, none of them seems to encompass all the functionalities described above. Therefore, the features of `Fitspy` make it an ideal tool for quickly fitting a few spectra through its GUI or for fitting several thousand of spectra (or more) by python batches, as can occur in the context of large-scale parametric studies (cf. @wafer).
+To the author's knowledge, although many open-source software applications are much more advanced in certain aspects mentioned, none of them seems to encompass all the functionalities described above. Therefore, the features of `Fitspy` make it an ideal tool for quickly fitting a few spectra through its GUI or for fitting several thousand of spectra (or more) by python batches, as can occur in the context of large-scale parametric studies [@wafer].
 
 
 # `Fitspy` workflow short description
@@ -57,7 +62,7 @@ To the author's knowledge, although many open-source software applications are m
 
[…]
