### 2024-01-05 f361a8e
minor typo corrections
