+++ fitspy/__init__.py
@@ -10,3 +10,5 @@ MODELS = {"Gaussian": gaussian,
 BKG_MODELS = ['None', 'Constant', 'Linear', 'Parabolic', 'Exponential']
 
 KEYS = ['x0', 'ampli', 'fwhm', 'fwhm_l', 'fwhm_r', 'alpha']
+
+VERSION = "2024.1"
