+++ doc/conf.py
@@ -1,0 +1,314 @@
+# -*- coding: utf-8 -*-
+#
+# PyStack3D documentation build configuration file, created by
+# sphinx-quickstart on Mon Oct 18 11:10:55 2010.
+#
+# This file is execfile()d with the current directory set to its containing dir.
+#
+# Note that not all possible configuration values are present in this
+# autogenerated file.
+#
+# All configuration values have a default; values that are commented out
+# serve to show the default.
+
+import sys
+from datetime import datetime
+
+sys.path.insert(0, '../')
+
+# from fitspy import Release
+import fitspy
+
+# If extensions (or modules to document with autodoc) are in another directory,
+# add these directories to sys.path here. If the directory is relative to the
+# documentation root, use os.path.abspath to make it absolute, like shown here.
+# sys.path.insert(0, os.path.abspath('.'))
+
+# -- General configuration -----------------------------------------------
+
+# If your documentation needs a minimal Sphinx version, state it here.
+# needs_sphinx = '1.0'
+
+# Add any Sphinx extension module names here, as strings. They can be extensions
+# coming with Sphinx (named 'sphinx.ext.*') or your custom ones.
+extensions = [
+    'sphinx.ext.autodoc',
+    'sphinx.ext.napoleon',
+    'sphinx.ext.intersphinx',
+    'sphinx.ext.mathjax',
+    'sphinx.ext.graphviz',
+    'sphinx.ext.autosummary',
+    'sphinx.ext.autosectionlabel',
+    'matplotlib.sphinxext.plot_directive',
+    # 'sphinx_toggleprompt',
+    # 'sphinxcontrib.towncrier',
+]
+
+linkcheck_ignore = [
+    "https://anaconda.org",  # 403 Client Error: Forbidden for url
+    "https://doi.org/10.1021/acs.nanolett.5b00449",
+    # 403 Client Error: Forbidden for url
+    "https://onlinelibrary.wiley.com",  # 403 Client Error: Forbidden for url
+]
+
+# linkcheck_exclude_documents = [
+#     'user_guide/io/*',
+#     'api/fitspy.io_plugins*',
+#     ]
+
+# Specify a standard user agent, as Sphinx default is blocked on some sites
+user_agent = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (" \
+             "KHTML, like Gecko) Chrome/108.0.0.0 Safari/537.36 " \
+             "Edg/108.0.1462.54"
+
+try:
+    import sphinxcontrib.spelling
+
+    extensions.append('sphinxcontrib.spelling')
+except BaseException:
+    pass
+# Add any paths that contain templates here, relative to this directory.
+templates_path = ['_templates']
+
+# The suffix of source filenames.
+source_suffix = '.rst'
+
+# The encoding of source files.
+# source_encoding = 'utf-8-sig'
+
+# The master toctree document.
+master_doc = 'index'
+
+# General information about the project.
+project = 'Fitspy'
+copyright = f'2023-{datetime.today().year}, The Fitspy development team'
+
+# The version info for the project you're documenting, acts as replacement for
+# |version| and |release|, also used in various other places throughout the
+# built documents.
+#
+# The short X.Y version.
+version = fitspy.VERSION
+# The full version, including alpha/beta/rc tags.
+release = fitspy.VERSION
+
+# The language for content autogenerated by Sphinx. Refer to documentation
+# for a list of supported languages.
+# language = None
+
+# There are two options for replacing |today|: either, you set today to some
+# non-false value, then it is used:
+# today = ''
+# Else, today_fmt is used as the format for a strftime call.
+# today_fmt = '%B %d, %Y'
+
[…]
