+++ doc/index.rst
@@ -1,0 +1,24 @@
+Welcome to Fitspy's documentation!
+====================================
+
+.. toctree::
+   :caption: User Guide
+
+   intro.rst
+   files_selection.rst
+   global_actions.rst
+   overall_settings.rst
+   baseline.rst
+   normalization.rst
+   fitting.rst
+   models.rst
+   figures.rst
+
+
+.. toctree::
+   :maxdepth: 2
+   :caption: API References
+
+   api/modules.rst
+
+
