+++ paper/paper.bib
@@ -15,13 +15,12 @@
 
 @software{PRISMA,
         author = {Flores, Eibar and Mozhzhukhina, Nataliia and Li, Xinyu and Norby, Poul and Matic, Aleksandar and Vegge, Tejs},
-        title = {PRISMA: A Robust and Intuitive Tool for High-Throughput Processing of Chemical Spectra**},
+        title = {PRISMA: A Robust and Intuitive Tool for High-Throughput Processing of Chemical Spectra},
         journal = {Chemistry–Methods},
         volume = {2},
         number = {10},
-        pages = {e202100094},
-        keywords = {electrochemical energy storage, high throughput analysis, in situ spectroscopy, operando spectroscopy, software},
-        doi = {https://doi.org/10.1002/cmtd.202100094},
+        keywords = {electrochemical energy storage, high throughput analysis, in situ spectroscopy, operando spectroscopy, software},
+        doi = {10.1002/cmtd.202100094},
         url = {https://chemistry-europe.onlinelibrary.wiley.com/doi/abs/10.1002/cmtd.202100094},
         eprint = {https://chemistry-europe.onlinelibrary.wiley.com/doi/pdf/10.1002/cmtd.202100094},
         abstract = {Abstract Here we introduce PRISMA, an open-source tool to rapidly analyse large numbers of spectra into meaningful spectroscopic trends. PRISMA follows a human-in-the-loop workflow, where the user interacts with an intuitive graphical user interface to control multiple steps in the spectrum analysis process: trimming, baseline correction, and peak fitting. The tool outputs the results in an easy-to-read csv format within seconds. We describe the functionalities implemented in PRISMA and test its capabilities with three experimental cases relevant to the study of Electrochemical energy storage and conversion devices: a temperature-dependent Raman measurement of phase transitions, a linear Raman mapping of a graphite composite electrode, and an operando X-ray diffraction experiment of lithium nickel oxide composite battery electrode. The case studies demonstrate the robustness of the app and its ability to unearth insightful trends in peak parameters, which are easier to represent, interpret and further analyse with more advanced techniques.},
@@ -90,10 +89,10 @@
                  Cadot, Stéphane and
                  Le Van-Jodin, Lucie},
         title = "{Full wafer-scale characterization method for 2D materials}",
- howpublished = {RamanFest 2023 conference},
+    publisher = {RamanFest 2023 conference},
      address  = {Paris},
          year = 2023,
         month = nov,
-       adsurl = {https://www.ramanfestconf.com/2023/Abstracts/2023_Meyer_Thibaut_52.pdf},
+       url = {https://www.ramanfestconf.com/2023/Abstracts/2023_Meyer_Thibaut_52.pdf},
       adsnote = {10th International Conference on Advanced Applied Raman Spectroscopy}
 }
