### 2024-01-08 269b2a7
add spectrum_coords()
