### 2024-01-09 #5868 @patquem
Hi @phibeck,
I am working on the article and a static HTML document (as previously requested by @maurov), which is taking a little bit of time. I plan to submit the final version of the article with my reply to the @maurov 's requests by the end of the week.
