### 2024-01-09 946e0e4
add xmin, xmax in load_profile()
