### 2024-01-11 d0a7e56
add gen_figures.py
