### 2024-01-12 faac952
add users model definition capability
