### 2024-01-13 c6325e8
BKG_MODEL as a dictionary with users bkg models
