+++ fitspy/__init__.py
@@ -1,4 +1,6 @@
 from pathlib import Path
+from lmfit.models import (ConstantModel, LinearModel, ParabolicModel,
+                          ExponentialModel)
 
 from fitspy.utils import get_func
 from fitspy.models import (gaussian, lorentzian,
@@ -14,8 +16,16 @@ MODELS = {"Gaussian": gaussian,
 
 PARAMS = ['x0', 'ampli', 'fwhm', 'fwhm_l', 'fwhm_r', 'alpha']
 
-BKG_MODELS = ['None', 'Constant', 'Linear', 'Parabolic', 'Exponential']
+BKG_MODELS = {'None': None,
+              'Constant': ConstantModel,
+              'Linear': LinearModel,
+              'Parabolic': ParabolicModel,
+              'Exponential': ExponentialModel}
 
 # add users models
-for model in get_func(Path.home() / "fitspy_users_models.txt"):
+for model in get_func(Path.home() / "fitspy_users_models.py"):
     MODELS.update({model[0]: model[1]})
+
+# add users BKG models
+for bkg_model in get_func(Path.home() / "fitspy_users_bkg_models.py"):
+    BKG_MODELS.update({bkg_model[0]: bkg_model[1]})
