### 2024-01-14 f8a7a67
simplify models.py and bkg_models.py interpretation
