### 2024-01-14 78c8e61
update bkg_model in plot
