### 2024-01-14 5702cdb
model creation with ExpressionModel object
