### 2024-01-14 eaa179a
manage ExpressionModel copy
