### 2024-01-14 807303c
add models from expressions in .txt
