### 2024-01-15 913808d
manage no spectrum in set_bkg_model()
