+++ fitspy/app/callbacks.py
@@ -587,10 +587,11 @@ class Callbacks:
         self.plot()
 
     def set_bkg_model(self):
-        self.current_spectrum.set_bkg_model(self.bkg_name.get())
-        self.current_spectrum.result_fit = None
-        self.tabview.update()
-        self.plot()
+        if self.current_spectrum is not None:
+            self.current_spectrum.set_bkg_model(self.bkg_name.get())
+            self.current_spectrum.result_fit = None
+            self.tabview.update()
+            self.plot()
 
     def update_fit_settings(self):
         """ Update fit settings """
