### 2024-01-15 6d4e961
manage bkg_model saving and reloading
