### 2024-01-15 b2f843b
add fitting.rst and release_notes.rst
