+++ doc/user_guide/fitting.rst
@@ -1,0 +1,26 @@
+Fitting
+=======
+
+The fitting processing is performed via the 'lmfit' package capabilities.
+
+
+Fitting parameters
+------------------
+
+(to be completed)
+
+* fitting model
+* number of iterations
+* weighting
+* multithreading
+
+convergence / minima
+
+
+Constraints settings
+--------------------
+
+(to be completed)
+
+* model parameters ranging
+* model parameters coupling
