+++ doc/release_notes/release_notes.rst
@@ -1,0 +1,20 @@
+Release Notes
+=============
+
+Version 2024.1 (january 15, 2024)
+---------------------------------
+
+New features:
+
+- to be completed
+
+
+Bug fixes/enhancements:
+
+- to be completed
+
+
+Version 2023.1 (july 10, 2023)
+------------------------------
+
+First release.
