### 2024-01-15 0f8d36d
article revisited
