+++ paper/paper.bib
@@ -88,11 +88,11 @@
                  Rolland, Emmanuel and
                  Cadot, Stéphane and
                  Le Van-Jodin, Lucie},
-        title = "{Full wafer-scale characterization method for 2D materials}",
+        title = "{Full wafer-scale characterization method for 2D materials.
+        10th International Conference on Advanced Applied Raman Spectroscopy}",
     publisher = {RamanFest 2023 conference},
      address  = {Paris},
          year = 2023,
         month = nov,
-       url = {https://www.ramanfestconf.com/2023/Abstracts/2023_Meyer_Thibaut_52.pdf},
-      adsnote = {10th International Conference on Advanced Applied Raman Spectroscopy}
+       url = {https://www.ramanfestconf.com/2023/Abstracts/2023_Meyer_Thibaut_52.pdf}
 }
