+++ paper/paper.md
@@ -8,7 +8,7 @@ tags:
   - fit
 authors:
   - name: Patrick Quéméré
-    orcid: 0009-0008-6936-1249 
+    orcid: 0009-0008-6936-1249
     affiliation: "1" # (Multiple affiliations must be quoted)
 affiliations:
   - name: Univ. Grenoble Alpes, CEA, Leti, F-38000 Grenoble, France
@@ -22,16 +22,14 @@ bibliography: paper.bib
 
 **`Fitspy`** is a dedicated tool to **fit** **sp**ectra in **py**thon.
 
-Spectrum decomposition (also known as fitting) is the choice of a linear combination of peak models to best represent an experimental spectrum. `Fitspy` currently has implementations for: Gaussian, Lorientzan, related asymmetric functions and the Pseudovoigt model. Spectrum decomposition corresponds to the resolution of a minimization problem, which can be subject to bounds and constraints on model parameters.
-
-The `Fitspy` GUI (see \autoref{fig:GUI}) is designed to be as simple and intuitive to use as possible.
+Spectrum decomposition (also known as fitting) is the choice of a linear combination of peak models to best represent an experimental spectrum. `Fitspy` currently has implementations for Gaussian, Lorentzian, related asymmetric functions and the Pseudovoigt model. User-defined models can also be added. For the fitting, the model parameters can be subject to bounds and constraints.
+From a practical point of view, the `Fitspy` GUI (see \autoref{fig:GUI}) is designed to be as simple and intuitive to use as possible.
 
 ![The `Fitpsy` GUI.\label{fig:GUI}](https://cea-metrocarac.github.io/fitspy/fitspy.png){width=85%}
 
 # Statement of need
 
-The analysis of spectra in many areas of physics, from materials characterisation to astrophysics, often requires their decomposition
-into more or less complex models to estimate the chemical composition of the subject being analysed.
+The analysis of spectra in many areas of physics, from materials characterisation to astrophysics, often requires their decomposition into more or less complex models to estimate the chemical composition of the subject being analysed.
 To carry out these decompositions, research communities can rely on spectral decomposition tools.
 While many open-source tools for spectral fitting exist, most of them have, however, been designed for specific application domains, offering a broad range of services beyond mere spectral fitting. Consequently, these tools can prove challenging to use, especially for less experienced individuals.
 
@@ -39,7 +37,7 @@ In the vein of generic tools like `Fityk` [@Fityk] or `PRISMA`  [@PRISMA], `Fits
 
 * **Agnostic Nature**: `Fitspy` is not tied to any specific physical quantity or database. It is designed to process spectra regardless of their x-support and y-intensity without any prior knowledge.
 
-* **Python Implementation**: `Fitspy` is coded in Python. As a result, spectra can be easily processed using Python scripts, catering to individuals with basic knowledge of the language.
+* **Python Implementation**: `Fitspy` is coded in Python. As a result, for individuals with basic knowledge of the language, spectra can be easily processed by scripts.
 
 * **2D Maps**: `Fitspy` has been designed to handle spectra derived from 2D acquisitions. Note that "2D" can encompass time or any other dimension. When dealing with 2D data, an interactive map in the `Fitspy` GUI allows users to locate and select spectra of interest easily.
 
@@ -51,16 +49,15 @@ In the vein of generic tools like `Fityk` [@Fityk] or `PRISMA`  [@PRISMA], `Fits
 
[…]
