### 2024-01-15 f4b683f
article revisited
