+++ .github/workflows/pytest.yml
@@ -1,6 +1,7 @@
 name: Python test
 
 on:
+  workflow_dispatch:
   push:
     branches: [ "main" ]
   pull_request:
