+++ setup.py
@@ -1,9 +1,7 @@
 from setuptools import setup, find_packages
-from fitspy import VERSION
 
 setup(
     name="fitspy",
-    version=VERSION,
     license='GPL v3',
     include_package_data=False,
     zip_safe=False,
