+++ pyproject.toml
@@ -5,3 +5,10 @@ build-backend = 'setuptools.build_meta'
 [tool.pytest.ini_options]
 minversion = "6.0"
 testpaths = ["tests"]
+
+[project]
+name = "fitspy"
+dynamic = ["version"]
+
+[tool.setuptools.dynamic]
+version = { attr = "fitspy.VERSION" }
