### 2024-01-15 829792a
project definition from pyproject.toml
