+++ pyproject.toml
@@ -14,7 +14,17 @@ keywords = ["Fitspy", "fit", "spectra", "spectrum", "map", "1D", "2D",
     "decomposition", "Gaussian", "Lorentzian", "Pseudovoigt", "GUI"]
 license = { text = "GPL v3" }
 classifiers = [
-    "Programming Language :: Python :: 3",
+    'Programming Language :: Python :: 3',
+    'Programming Language :: Python :: 3.7',
+    'Programming Language :: Python :: 3.8',
+    'Programming Language :: Python :: 3.9',
+    'Programming Language :: Python :: 3.10',
+    'Programming Language :: Python :: 3.11',
+    'License :: OSI Approved :: GNU General Public License v3 (GPLv3)',
+    'Operating System :: Microsoft :: Windows',
+    'Operating System :: POSIX :: Linux',
+    'Operating System :: MacOS :: MacOS X',
+    'Environment :: Console',
 ]
 dependencies = [
     "numpy",
@@ -25,7 +35,6 @@ dependencies = [
     "parse",
     "pywin32; platform_system == 'Windows'",
 ]
-dynamic = ["version"]
 
 [project.scripts]
 fitspy = "fitspy.app.gui:fitspy_launcher"
@@ -38,7 +47,7 @@ version = { attr = "fitspy.VERSION" }
 
 [tool.setuptools.packages.find]
 where = ["."]
-include = ["fitspy*"]  # alternatively: `exclude = ["additional*"]`
+include = ["fitspy*"]
 namespaces = false
 
 [tool.pytest.ini_options]
