### 2024-01-16 #5868 @patquem
Hello @phibeck, @maurov and @FCMeng,
As agreed, regarding the article, the 'statement of needs' section has been revised, and an illustration corresponding to a use case has been added.
The documentation is now in the form of a static HTML Sphinx doc.
Fitspy has also undergone some improvements, including the ability for users to input their own models.
A new version of fitspy (v2024.1) has just been deposited on PyPI today.
Regarding the issues raised by @maurov, not all of them have been addressed yet.
Hoping that the project's status and the article meet your expectations to authorize the upcoming publication in JOSS.
