### 2024-01-16 394bdeb
remove extra tabs in 2D map example
