+++ doc/user_guide/input_data.rst
@@ -50,7 +50,7 @@ Note that (X) and (Y) are not necessary related to spacial dimensions. They can 
 
 Example of 2D-map input data::
 
-    \t  \t  958.468750	956.974609  955.480469	953.984375	952.490234	...
+    \t  \t  958.468750  956.974609  955.480469  953.984375  952.490234  ...
     0.  0.  307.873749  292.814697  327.923798  307.819580  314.483063  ...
     1.  0.  308.425364  291.142875  320.452365  305.234865  312.325687  ...
     2.  0.  306.325648  291.235672  324.834579  306.378541  312.387649  ...
