### 2024-01-18 c4b087d
manage linewidth in plot()
