### 2024-01-18 65a25f6
revisit baseline attributes
