### 2024-01-18 80bc760
add (optional) baseline display
