### 2024-01-18 86ef8e0
display the legend
