+++ fitspy/spectrum.py
@@ -48,6 +48,12 @@ class Spectrum:
         List of labels associated to the models. Default is ['1', '2', '3', ...]
     models_index: itertools.count
         Counter used for models indexing when creating a lmfit.Model
+    baseline: Baseline object
+        Baseline associated to the spectrum (to subract)
+    baseline_history: list of list - DEPRECATED from v2024.2
+        Concatenation list of all baselines applied to the spectrum. Each item
+        of the list consist in the max order of the baseline polynom and the
+        (x, y) baseline points coordinates.
     bkg_model: lmfit.Model
         Background model to fit with the composite peaks models, among :
         [None, 'ConstantModel', 'LinearModel', 'ParabolicModel',
@@ -64,12 +70,6 @@ class Spectrum:
         Default is 200.
     result_fit: lmfit.ModelResult
         Object resulting from lmfit fitting
-    baseline: Baseline object
-        Baseline associated to the spectrum (to be removed)
-    baseline_history: list of list
-        Concatenation list of all baselines applied to the spectrum. Each item
-        of the list consist in the max order of the baseline polynome and the
-        (x, y) baseline points coordinates
     """
 
     def __init__(self):
@@ -86,24 +86,24 @@ class Spectrum:
         self.peaks = None
         self.peaks_params = {'distance': 20, 'prominence': None,
                              'width': None, 'height': None, 'threshold': None}
+        self.baseline = BaseLine()
+        self.bkg_model = None
         self.models = []
         self.models_labels = []
         self.models_index = itertools.count(start=1)
-        self.bkg_model = None
         self.fit_method = 'leastsq'
         self.fit_negative = False
         self.max_ite = 200
         self.result_fit = None
 
-        self.baseline = BaseLine()
-        self.baseline_history = []
-
     def set_attributes(self, dict_attrs, **fit_kwargs):
         """Set attributes from a dictionary (obtained from a .json reloading)"""
 
-        for key, val in dict_attrs.items():
-            if key not in ['models', 'bkg_model']:
-                setattr(self, key, val)
+        keys = dict_attrs.keys()
+
+        for key in vars(self).keys():
+            if key in keys:
+                setattr(self, key, dict_attrs[key])
 
         for key in ['fit_method', 'fit_negative', 'max_ite']:
             if key in fit_kwargs:
@@ -123,7 +123,7 @@ class Spectrum:
                 model.param_hints = param_hints
                 self.models.append(model)
 
-        if 'bkg_model' in dict_attrs.keys() and dict_attrs['bkg_model']:
+        if 'bkg_model' in keys and dict_attrs['bkg_model']:
             model_name, param_hints = list(dict_attrs['bkg_model'].items())[0]
             model = BKG_MODELS[model_name]
             if isinstance(model, ExpressionModel):
@@ -135,21 +135,45 @@ class Spectrum:
             self.bkg_model = model
             self.bkg_model.name2 = model_name
             self.bkg_model.param_hints = param_hints
+            print(self.bkg_model)
 
-        # to make 'old' models still working
-        if "models_labels" not in dict_attrs.keys() or \
-                len(dict_attrs["models_labels"]) == 0:
+        if 'baseline' in keys:
+            self.baseline = BaseLine()
+            for key, val in vars(self.baseline).items():
+                if key in dict_attrs['baseline'].keys():
+                    setattr(self.baseline, key, dict_attrs['baseline'][key])
+
+        # COMPATIBILITY with 'old' models
+        #################################
+
+        if "models_labels" not in keys or len(dict_attrs["models_labels"]) == 0:
             self.models_labels = list(map(str, range(1, len(self.models) + 1)))
+
+        if "baseline_history" in keys:
[…]
