### 2024-01-19 eaa4823
display the model parameters whatever the fit status
