### 2024-01-19 0fbad7e
manage result_fit when saving and reloading
