### 2024-01-19 a2067e9
bugfix: reinitialize norm_mode
