### 2024-01-19 004c3ea
manage WindowsPath in hsorted()
