### 2024-01-20 8772030
check 'models' in dict_attrs.keys()
