### 2024-01-20 3cf2b2a
use get_objects() to find spectrum from fname
