### 2024-01-22 129ffc0
deprecated argument 'model' in appli_model()
