### 2024-01-22 74c2db6
save_params() and save_stats() moved from spectra.py to spectrum.py
