### 2024-01-24 f3eed9a
bkg_model correctly taken into account in fit plotting
