### 2024-01-24 382c2ce
Manage ExpressionModel when creating model
