### 2024-01-24 a4bcd40
manage ExpressionModel in fit_mp
