### 2024-01-24 9dbb3cf
destroy previuos spectra_map.frame when reloading
