### 2024-01-24 97ea49f
revisit fit() using apply_model()
