### 2024-01-24 51459ec
define spectrum.result_fit as an object
