### 2024-01-25 a171b59
factorize fit() by using apply_model()
