### 2024-01-25 eb6746b
bug fixed related to fnames passed by fit() to apply_mode()
