### 2024-01-25 6210ada
add lmfit reference
