### 2024-01-26 f2a6bdb
add load_models_from_txt() and load_models_from_py() in utils.py
