### 2024-01-26 d4ee38a
add thread.join() to wait process ending
