### 2024-01-26 60d4574
add ex_gui_users_defined_models.py and related test
