+++ examples/data/2D_maps/models.txt
@@ -1,0 +1,1 @@
+Lorentzian_user = ampli * fwhm ** 2 / (4 * ((x - x0) ** 2 + fwhm ** 2 / 4))
