+++ examples/data/2D_maps/bkg_models.py
@@ -4,8 +4,8 @@ module description
 from fitspy import BKG_MODELS
 
 
-def linear_user(x, slope, constant):
-    return slope * x + constant
+def linear_user(x, slope, intercept):
+    return slope * x + intercept
 
 
 BKG_MODELS.update({'Linear_user': linear_user})
