+++ examples/data/2D_maps/bkg_models.txt
@@ -1,1 +1,1 @@
-Linear_user = slope * x + constant
+Linear_user = slope * x + intercept
