### 2024-01-29 f4fa450
consider fit_report as a spectrum.resul_fit attribute
