### 2024-01-29 f72d9e7
bugfix: bkg_name was not correctly updated
