### 2024-01-29 1276173
consider fit_report as a spectrum.resul_fit attribute
