### 2024-01-29 c7c5db8
bugfix (forgetting): define spectrum.result_fit as an object
