+++ fitspy/app/callbacks.py
@@ -734,7 +734,7 @@ class Callbacks:
             spectrum.x = spectrum.x0.copy()
             spectrum.y = spectrum.y0.copy()
             spectrum.norm_mode = None
-            spectrum.result_fit = None
+            spectrum.result_fit = lambda: None
             spectrum.remove_models()
             spectrum.baseline.points = [[], []]
             spectrum.baseline.is_subtracted = False
