### 2024-01-29 d9c2141
bugfix: color panel management with result_fit.success
