### 2024-01-29 fa907ee
bugfix: apply_model with progressbar should be in the main thread
