+++ fitspy/spectra.py
@@ -11,38 +11,33 @@ from threading import Thread
 from multiprocessing import Queue
 import matplotlib.pyplot as plt
 from lmfit.models import ExpressionModel
+import dill
 
 from fitspy.utils import fileparts, save_to_json, load_from_json
 from fitspy.spectrum import Spectrum
-from fitspy import MODELS
+from fitspy import MODELS, MODELS_NAMES
 
 
 def fit(params):
     """ Fitting function used in multiprocessing """
-    x, y, models, method, fit_negative, max_ite = params
+    x, y, models_, method, fit_negative, max_ite = params
+
+    models = []
+    for model_ in models_:
+        if isinstance(model_, bytes):
+            models.append(dill.loads(model_))
+        else:
+            models.append(model_)
+
     spectrum = Spectrum()
     spectrum.x = x
     spectrum.y = y
-
-    models_ = []
-    for model in models:
-        if isinstance(model, dict):
-            model_ = ExpressionModel(model['expr'], independent_vars=['x'])
-            model_.__name__ = model['name']
-            model_.param_hints = deepcopy(model['param_hints'])
-            model_.make_params()
-            param_names = model_.param_names.copy()
-            model_.prefix = model['prefix']  # -> make model.param_names = []
-            for name in param_names:  # reassign model.param_names with prefix
-                model_.param_names.append(model['prefix'] + name)
-            models_.append(model_)
-        else:
-            models_.append(model)
-
-    spectrum.models = models_
+    spectrum.models = models
     spectrum.fit(fit_method=method, fit_negative=fit_negative, max_ite=max_ite)
     shared_queue.put(1)
-    return spectrum.result_fit.values, spectrum.result_fit.success
+
+    result_fit = spectrum.result_fit
+    return result_fit.values, result_fit.success, result_fit.fit_report
 
 
 def initializer(queue_incr):
@@ -51,31 +46,32 @@ def initializer(queue_incr):
     shared_queue = queue_incr
 
 
-def fit_mp(spectra, models, bkg_model,
-           fit_method, fit_negative, max_ite, ncpus, queue_incr):
+def fit_mp(spectra, ncpus, queue_incr):
     """ Multiprocessing fit function applied to spectra """
 
     ncpus = ncpus or os.cpu_count()
     ncpus = min(ncpus, os.cpu_count())
 
-    def picklable_model(model):
-        if isinstance(model, ExpressionModel):
-            return {'expr': model.expr,
-                    'name': model.__name__,
-                    'prefix': model.prefix,
-                    'param_hints': model.param_hints}
+    spectrum = spectra[0]
+    models_ = []
+    for model in spectrum.models:
+        if model.name2 not in MODELS_NAMES:
+            models_.append(dill.dumps(model))
         else:
-            return model
+            models_.append(model)
+    if spectrum.bkg_model is not None:
+        if spectrum.bkg_model.name2 not in MODELS_NAMES:
+            models_.append(dill.dumps(spectrum.bkg_model))
+        else:
+            models_.append(spectrum.bkg_model)
+    fit_method = spectrum.fit_method
+    fit_negative = spectrum.fit_negative
+    max_ite = spectrum.max_ite
 
     args = []
     for spectrum in spectra:
         x, y = spectrum.x, spectrum.y
-        models_ = []
-        for model in models:
-            models_.append(picklable_model(model))
-        models_.append(picklable_model(bkg_model))
-        args.append((x, y, models_,
-                     fit_method, fit_negative, max_ite))
+        args.append((x, y, models_, fit_method, fit_negative, max_ite))
 
     with ProcessPoolExecutor(initializer=initializer,
                              initargs=(queue_incr,),
[…]
