+++ fitspy/__init__.py
@@ -12,13 +12,13 @@ VERSION = "2024.2beta"
 FITSPY_DIR = Path.home() / "Fitspy"
 SETTINGS_FNAME = FITSPY_DIR / "settings.json"
 
-MODELS = {"Gaussian": gaussian,
-          "Lorentzian": lorentzian,
-          "PseudoVoigt": pseudovoigt,
-          "GaussianAsym": gaussian_asym,
-          "LorentzianAsym": lorentzian_asym}
+PEAK_MODELS = {"Gaussian": gaussian,
+               "Lorentzian": lorentzian,
+               "PseudoVoigt": pseudovoigt,
+               "GaussianAsym": gaussian_asym,
+               "LorentzianAsym": lorentzian_asym}
 
-PARAMS = ['x0', 'ampli', 'fwhm', 'fwhm_l', 'fwhm_r', 'alpha']
+PEAK_PARAMS = ['x0', 'ampli', 'fwhm', 'fwhm_l', 'fwhm_r', 'alpha']
 
 BKG_MODELS = {'None': None,
               'Constant': ConstantModel,
@@ -26,7 +26,7 @@ BKG_MODELS = {'None': None,
               'Parabolic': ParabolicModel,
               'Exponential': ExponentialModel}
 
-MODELS_NAMES = list(MODELS.keys()) + list(BKG_MODELS.keys())
+MODELS_NAMES = list(PEAK_MODELS.keys()) + list(BKG_MODELS.keys())
 
 # create FITSPY_DIR if not exists
 Path.mkdir(FITSPY_DIR, exist_ok=True)
@@ -37,9 +37,9 @@ if fname.exists():
     shutil.move(fname, SETTINGS_FNAME)
 
 # add users models from '.txt' file
-load_models_from_txt(FITSPY_DIR / "models.txt", MODELS)
+load_models_from_txt(FITSPY_DIR / "peak_models.txt", PEAK_MODELS)
 load_models_from_txt(FITSPY_DIR / "bkg_models.txt", BKG_MODELS)
 
 # add users models from '.py' file
-load_models_from_py(FITSPY_DIR / "models.py")
+load_models_from_py(FITSPY_DIR / "peak_models.py")
 load_models_from_py(FITSPY_DIR / "bkg_models.py")
