+++ examples/ex_gui_users_defined_models.py
@@ -6,7 +6,7 @@ from pathlib import Path
 
 from fitspy.app.gui import Appli
 from fitspy.utils import load_models_from_txt, load_models_from_py
-from fitspy import MODELS, BKG_MODELS
+from fitspy import PEAK_MODELS, BKG_MODELS
 
 DATA = Path(__file__).parent / "data"
 DIRNAME = DATA / "2D_maps"
@@ -14,14 +14,14 @@ DIRNAME = DATA / "2D_maps"
 
 def ex_gui_users_models_from_txt(ncpus=1, dirname_res=None):
     """ Example using users models defined from literal expression in a .txt """
-    load_models_from_txt(DIRNAME / "models.txt", MODELS)
+    load_models_from_txt(DIRNAME / "peak_models.txt", PEAK_MODELS)
     load_models_from_txt(DIRNAME / "bkg_models.txt", BKG_MODELS)
     ex_gui_users_models(ncpus=ncpus, dirname_res=dirname_res)
 
 
 def ex_gui_users_models_from_py(ncpus=1, dirname_res=None):
     """ Example using users models defined from functions in a .py """
-    load_models_from_py(DIRNAME / "models.py")
+    load_models_from_py(DIRNAME / "peak_models.py")
     load_models_from_py(DIRNAME / "bkg_models.py")
     ex_gui_users_models(ncpus=ncpus, dirname_res=dirname_res)
 
@@ -37,6 +37,7 @@ def ex_gui_users_models(ncpus=1, dirname_res=None):
     appli.add_items(fnames=[DIRNAME / 'ordered_map.txt'])
     appli.load_model(fname_json=fname_json)
     appli.apply_model(fnames=appli.spectra.fnames[:5])
+    # appli.apply_model() # apply model to all the spectra
 
     # save and destroy for pytest
     if dirname_res is not None:
