### 2024-01-29 eba831b
change 'model' and 'dict_attrs' names to 'model_dict'
