### 2024-01-29 04f599a
rename some Spectrum attributes (peaks -> attractors, models -> peak_models)
