+++ fitspy/spectrum.py
@@ -22,7 +22,7 @@ from fitspy import PEAK_MODELS, PEAK_PARAMS, BKG_MODELS
 
 
 def create_model(model, model_name, prefix=None):
-    """ Create a 'model' (peak_model or 'bkg_model') object """
+    """ Return a 'model' (peak_model or 'bkg_model') object """
     if isinstance(model, ExpressionModel):
         model = ExpressionModel(model.expr, independent_vars=['x'])
         model.__name__ = model_name
@@ -60,17 +60,11 @@ class Spectrum:
         Arrays related to spectrum raw support and intensity (resp.)
     x, y: numpy.array((n))
         Arrays related to spectrum modified support and intensity (resp.)
-    peaks: list of ints
-        Indices of peaks
-    peaks_params: dict
-        Dictionary passed to scipy.signal.find_peaks for peaks determination.
-        Could be a dictionary with Tkinter.Variable as values.
-    models: list of lmfit.Model
-        List of the lmfit models
-    models_labels: list of str
-        List of labels associated to the models. Default is ['1', '2', '3', ...]
-    models_index: itertools.count
-        Counter used for models indexing when creating a lmfit.Model
+    attractors: list of ints
+        Indices of attractors
+    attractors_params: dict
+        Dictionary passed to scipy.signal.find_peaks for attractors
+        determination. Could be a dictionary with Tkinter.Variable as values.
     baseline: Baseline object
         Baseline associated to the spectrum (to subract)
     baseline_history: list of list - DEPRECATED from v2024.2
@@ -81,6 +75,12 @@ class Spectrum:
         Background model to fit with the composite peaks models, among :
         [None, 'ConstantModel', 'LinearModel', 'ParabolicModel',
         'ExponentialModel']
+    peak_models: list of lmfit.Model
+        List of peak models
+    peak_labels: list of str
+        List of labels associated to the peak models. Default is ['1', '2', ...]
+    peak_index: itertools.count
+        Counter used for peak models indexing when creating a lmfit.Model
     fit_method: str
         Method used for fitting. See lmfit.Model.fit().
         Default method is 'leastsq'.
@@ -110,14 +110,15 @@ class Spectrum:
         self.y0 = None
         self.x = None
         self.y = None
-        self.peaks = None
-        self.peaks_params = {'distance': 20, 'prominence': None,
-                             'width': None, 'height': None, 'threshold': None}
+        self.attractors = None
+        self.attractors_params = {'distance': 20, 'prominence': None,
+                                  'width': None, 'height': None,
+                                  'threshold': None}
         self.baseline = BaseLine()
         self.bkg_model = None
-        self.models = []
-        self.models_labels = []
-        self.models_index = itertools.count(start=1)
+        self.peak_models = []
+        self.peak_labels = []
+        self.peak_index = itertools.count(start=1)
         self.fit_method = 'leastsq'
         self.fit_negative = False
         self.max_ite = 200
@@ -128,6 +129,18 @@ class Spectrum:
 
         keys = model_dict.keys()
 
+        # compatibility with 'old' key names
+        if 'peaks' in keys:
+            model_dict['attractors'] = model_dict.pop('peaks')
+        if 'peaks_params' in keys:
+            model_dict['attractors_params'] = model_dict.pop('peaks_params')
+        if 'models' in keys:
+            model_dict['peak_models'] = model_dict.pop('models')
+        if 'models_labels' in keys:
+            model_dict['peak_labels'] = model_dict.pop('models_labels')
+        if 'models_index' in keys:
+            model_dict['peak_index'] = model_dict.pop('models_index')
+
         for key in vars(self).keys():
             if key in keys:
                 setattr(self, key, model_dict[key])
[…]
