+++ fitspy/app/callbacks.py
@@ -142,7 +142,7 @@ class Callbacks:
         self.fr_baseline.disable()
 
     def save_results(self, dirname_res=None):
-        """ Save all spectra (peaks) results in .csv files """
+        """ Save all results (peaks parameters) in .csv files """
         if dirname_res is None:
             dirname_res = fd.askdirectory(title='Select directory')
         else:
@@ -227,7 +227,7 @@ class Callbacks:
         for fname in fnames:
             for spectrum in self.spectra:
                 if spectrum.fname == fname:
-                    if len(spectrum.models) == 0:
+                    if len(spectrum.peak_models) == 0:
                         no_models.append(os.path.basename(fname))
                     break
 
@@ -301,7 +301,7 @@ class Callbacks:
             self.update()
 
     def plot(self):
-        """ Plot baseline and spectrum models after 'ax' clearing """
+        """ Plot baseline and peak models after 'ax' clearing """
         if not self.show_plot or self.current_spectrum is None:
             return
 
@@ -332,12 +332,12 @@ class Callbacks:
         self.ax.set_ylabel(fig_settings['y_label'].get(), fontsize=18)
 
         if fig_settings['plot_fit'].get() == 'On':
-            show_peaks = self.attractors.get()
+            show_attractors = self.attractors.get()
             show_neg_values = fig_settings['plot_negative_values'].get() == 'On'
             show_baseline = fig_settings['plot_baseline'].get() == 'On'
             show_background = fig_settings['plot_background'].get() == 'On'
             self.lines = spectrum.plot(self.ax,
-                                       show_peaks=show_peaks,
+                                       show_attractors=show_attractors,
                                        show_negative_values=show_neg_values,
                                        show_baseline=show_baseline,
                                        show_background=show_background)
@@ -356,13 +356,13 @@ class Callbacks:
             def annotate_params(i, color='k'):
                 """ Annotate figure with fit parameters """
                 if not line_bkg_visible:
-                    model = spectrum.models[i]
+                    model = spectrum.peak_models[i]
                     x0 = model.param_hints['x0']['value']
                 elif i == 0:
                     model = spectrum.bkg_model
                     x0 = 0.5 * (x[0] + x[-1])
                 else:
-                    model = spectrum.models[i - 1]
+                    model = spectrum.peak_models[i - 1]
                     x0 = model.param_hints['x0']['value']
 
                 y0 = model.eval(model.make_params(), x=x0)
@@ -396,10 +396,10 @@ class Callbacks:
 
         if fig_settings['show_peaks_labels'].get() == 'On':
             dy = 0.02 * spectrum.y.max()
-            for i, label in enumerate(spectrum.models_labels):
+            for i, label in enumerate(spectrum.peak_labels):
                 if label == '':
                     continue
-                model = spectrum.models[i]
+                model = spectrum.peak_models[i]
                 x0 = model.param_hints['x0']['value']
                 y = spectrum.y[closest_index(spectrum.x, x0)]
                 xy = (x0, y + dy)
@@ -536,17 +536,17 @@ class Callbacks:
         self.current_spectrum.baseline.points = [[], []]
         self.plot()
 
-    def update_attractors(self):
-        """ Update attractors """
-        if self.attractors.get():
-            self.current_spectrum.peaks_calculation()
-        self.plot()
+    # def update_attractors(self):
+    #     """ Update attractors """
+    #     if self.attractors.get():
+    #         self.current_spectrum.attractors_calculation()
+    #     self.plot()
 
     def update_attractors_settings(self):
[…]
