### 2024-01-29 cc9a269
manage default values for attractors_params
