### 2024-01-29 ce0e322
pass x0 to add_peak_model()
