+++ doc/README.md
@@ -1,197 +1,0 @@
-# Fitspy Documentation
-
-## Introduction
-
-The main class of Fitspy is [`Spectra`](../fitspy/spectra.py) (inherited
-from `list`) that contains a list of  [`Spectrum`](../fitspy/spectra.py)
-objects.
-
-For each of these [`Spectrum`](../fitspy/spectra.py) objects, the corresponding (x, y) profile is associated.
-
-Additionally, 2D maps of spectra (named `SpectraMap`) can be provided as input according to a dedicated format. They consist in a set of `Spectrum` with associated grid coordinates (see below).
-
-![](_static/spectra.png)
-
-The spectra processing are based on  `baselines` definition (optional) and `peaks` models definition (Gaussian, Lorentzian, ...) used to fit the spectra profiles.
-
-**Important: In terms of usage, to avoid confusion, it is recommended to address spectra processing from the top to the bottom frames as depicted hereafter**.
-
-## Preliminary remarks
-
-**Axis limits**<br>
-To facilitate the comparison between spectra visually, the (x, y) axis limits in the figure are set based on the first displayed spectrum.<br>
-Subsequently, these limits are applied to all the following spectra.<br>
-To readjust the axis limits to the current spectrum click on <img src="_static/home.png" width="20"> of the navigation toolbar (auto-rescaling).
-
-**Mouse management**<br>
-Mouse click buttons are used both with the navigation toolbar (for zooming and translating) and for defining baseline points or peaks points.<br>
-The activation/deactivation of the mouse associated with these last 2 modes (baseline and peaks selection) requires the corresponding frames to be "clicked" beforehand.
-
-## Files selection
-
-The files selection is realized thanks to the following widget:
-
-![](_static/files_selector.png)
-
-Two types of data can be interpreted.
-
-* The first type consists of input data files containing 2 columns representing the support (x) and spectrum intensity (y) respectively.
-  The separators between the columns can be tabulation `\t`, comma `,`, semicolon `;` or space ` `. Note that the first row is skipped and the (x, y) rows can be unordered.
-
-
-* The second type is associated with **2D-map** input data. the first line of the file is related to the support (x) of the spectra and starts with 2 tabulations `\t`. The subsequent rows are related to the spectra intensity at the 'pixel' coordinates (X), (Y) defined in the first 2 columns.
-  Note that the number of intensity values must be consistent with the support (x) defined in the first row. X-Y spectra can be defined in a partial and unordered 2D grid.
-
-  Once loaded, a 2D-map is expanded. The associated spectra appear in the files selection widget to according to the 2D-map file basename and the (X) and (Y) spectrum coordinates in the grid.
-  A 2D-map figure is also opened to facilitate interaction with the cursor selection of the files selector widget.
-
-<p align="center" width="100%">
-  interactive 2D-maps figures resulting from a <a href="../examples/data/2D_maps/ordered_map.txt">fully defined ordered acquisition grid</a> (left) and a <a href="../examples/data/2D_maps/unordered_map.txt">partially unordered one</a> (right) <br> <br>
-  <img align="justify" width="25%" src=_static/2d-map.png>
-  <img align="justify" width="25%" src=_static/2d-map_unordered.png>
-</p>
-
-## Global actions
-
-![](_static/global_actions.png)
-
-`Show All` enables a global display of all the spectra. When clicking on the canvas with the mouse, the nearest spectra are highlighted (up to a maximum of 10 spectra).
-
-`Auto eval` and `Auto eval All`  perform automatic evaluation of the baseline and peak positions and conduct fitting on the selected spectra or all the spectra (respectively).
-
[…]
