### 2024-01-30 3c1ccf1
bugfix: peak_index has to be reinitialize to 1 with set_attributes()
