### 2024-01-31 5bcdaa6
change result_fit.fit_report in result_fit.report
