### 2024-01-31 e22f282
Attempt of a method limiting the drill writing/loading of the models (aborted)
