+++ fitspy/utils_mp.py
@@ -1,34 +1,35 @@
 """
 utilities functions related to multiprocessing
+
+notes:
+The strategy (see commented lines below) of passing the models to the workers
+once instead of duplicating them for each spectrum turned out to be slightly
+  more costly in terms of CPU time finally (?).
 """
 import os
 from concurrent.futures import ProcessPoolExecutor
 import dill
 
 from fitspy.spectrum import Spectrum
-from fitspy import MODELS_NAMES, PEAK_MODELS
 
 
 def fit(params):
     """ Fitting function used in multiprocessing """
     x, y, models_, method, fit_negative, max_ite = params
 
-    models = []  # all peak_models and bkg_model have been put in 'models_'
+    models = []
     for model_ in models_:
-        if isinstance(model_, bytes):
-            models.append(dill.loads(model_))
-        else:
-            models.append(model_)
+        models.append(dill.loads(model_))
 
     spectrum = Spectrum()
     spectrum.x = x
     spectrum.y = y
-    spectrum.peak_models = models
+    spectrum.peak_models = models  # MODELS = peak_models + bkg_model
     spectrum.fit(fit_method=method, fit_negative=fit_negative, max_ite=max_ite)
     shared_queue.put(1)
 
     result_fit = spectrum.result_fit
-    return result_fit.values, result_fit.success, result_fit.fit_report
+    return result_fit.values, result_fit.success, result_fit.report
 
 
 def initializer(queue_incr):
@@ -47,17 +48,12 @@ def fit_mp(spectra, ncpus, queue_incr):
     fit_method = spectrum.fit_method
     fit_negative = spectrum.fit_negative
     max_ite = spectrum.max_ite
+
     models_ = []  # all peak_models and bkg_model are put in a single 'models_'
     for peak_model in spectrum.peak_models:
-        if peak_model.name2 not in MODELS_NAMES:
-            models_.append(dill.dumps(peak_model))
-        else:
-            models_.append(peak_model)
+        models_.append(dill.dumps(peak_model))
     if spectrum.bkg_model is not None:
-        if spectrum.bkg_model.name2 not in MODELS_NAMES:
-            models_.append(dill.dumps(spectrum.bkg_model))
-        else:
-            models_.append(spectrum.bkg_model)
+        models_.append(dill.dumps(spectrum.bkg_model))
 
     args = []
     for spectrum in spectra:
@@ -69,17 +65,103 @@ def fit_mp(spectra, ncpus, queue_incr):
                              max_workers=ncpus) as executor:
         results = tuple(executor.map(fit, args))
 
-    # dictionary of custom function names and definitions
-    funcdefs = {}
-    for val in PEAK_MODELS.values():
-        funcdefs[val.__name__] = val
-
-    for (values, success, fit_report), spectrum in zip(results, spectra):
+    for (values, success, report), spectrum in zip(results, spectra):
         spectrum.result_fit.success = success
-        spectrum.result_fit.fit_report = fit_report
+        spectrum.result_fit.report = report
         for peak_model in spectrum.peak_models:
             for key in peak_model.param_names:
                 peak_model.set_param_hint(key[4:], value=values[key])
         if spectrum.bkg_model is not None:
             for key in spectrum.bkg_model.param_names:
                 spectrum.bkg_model.set_param_hint(key, value=values[key])
+
+# import os
+# from concurrent.futures import ProcessPoolExecutor
+# from copy import deepcopy
+# import dill
+#
+# from fitspy.spectrum import Spectrum
+# from fitspy import PEAK_MODELS
+#
+#
+# def fit(params):
+#     """ Fitting function used in multiprocessing """
[…]
