### 2024-01-31 3f0ffc0
add option to reinitialize 'ampli' and 'fwhm' guess in fit()
