### 2024-02-01 17d7ee7
change baseline_distance from DoubleVar to IntVar to be in settings
