### 2024-02-01 72fe8b8
check comp_model is not None when applying reinit_guess
