### 2024-02-01 d7026b7
bugfix: use self.fnames instead of self.all in apply_model()
