### 2024-02-02 1710483
manage ncpus in Spectra.apply_model() instead of utils_mp.fi_mp()
