### 2024-02-02 2f94081
manage grid range in the case of single X or Y values
