+++ examples/ex_nogui_model_reconstruction_from_params.py
@@ -1,5 +1,5 @@
 """
-Example of spectrum model reconstruction from .json and .csv parameters
+Example of spectrum peak models reconstruction from .json and .csv parameters
 """
 from pathlib import Path
 import pandas as pd
@@ -11,16 +11,16 @@ from fitspy.spectra import Spectra, Spectrum
 DATA = Path(__file__).parent / "data"
 
 
-def plot_spectrum_models(spectrum, title=None):
-    """ Plot 'spectrum' model """
+def plot_spectrum_peak_models(spectrum, title=None):
+    """ Plot 'spectrum' peak models """
 
-    # x-support for spectrum models evaluation
+    # x-support for spectrum peak models evaluation
     x = np.linspace(100, 800., 200)
 
     plt.figure()
     y_sum = np.zeros_like(x)
-    for model in spectrum.models:
-        y = model.eval(model.make_params(), x=x)
+    for peak_model in spectrum.peak_models:
+        y = peak_model.eval(peak_model.make_params(), x=x)
         y_sum += y
         plt.plot(x, y)
     plt.plot(x, y_sum, 'b')
@@ -35,7 +35,7 @@ def model_reconstruction_from_json():
     # considering the first spectrum model
     spectrum = Spectra.load(fname_json)[0]
 
-    plot_spectrum_models(spectrum, title='from .json')
+    plot_spectrum_peak_models(spectrum, title='from .json')
 
 
 def model_reconstruction_from_csv():
@@ -44,16 +44,16 @@ def model_reconstruction_from_csv():
     fname_csv = DATA / 'spectra_2' / 'InP-1_42-P21.csv'
 
     dfr = pd.read_csv(fname_csv, sep=';')
-    models = []
+    peak_models = []
     for i in dfr.index:
         params = list(dfr.iloc[i, 1:])
-        model = Spectrum.create_model(i, *params)
-        models.append(model)
+        peak_model = Spectrum.create_peak_model(i, *params)
+        peak_models.append(peak_model)
 
     spectrum = Spectrum()
-    spectrum.models = models
+    spectrum.peak_models = peak_models
 
-    plot_spectrum_models(spectrum, title='from .csv')
+    plot_spectrum_peak_models(spectrum, title='from .csv')
 
 
 if __name__ == "__main__":
