### 2024-02-02 cebcc61
put 'fit_method', 'fit_negative' and 'max_ite' in spectrum.fit_params dictionary
