### 2024-02-02 9f6f94e
pass peak_models and bkg_models from dill.dumps/dill.loads
