### 2024-02-02 73a5a8c
return all spectrum.result_fit with dill.dumps/dill.loads
