### 2024-02-02 92cf804
add epsilon in the Lorentzian denominator
