### 2024-02-02 60b5f39
add noise level management in fit()
