+++ fitspy/spectrum.py
@@ -9,6 +9,7 @@ from copy import deepcopy
 import numpy as np
 import pandas as pd
 from scipy.signal import find_peaks
+from scipy.ndimage import uniform_filter1d
 from lmfit import Model, fit_report
 from lmfit.model import ModelResult
 from lmfit.models import ConstantModel, LinearModel, ParabolicModel, \
@@ -23,7 +24,8 @@ from fitspy import PEAK_MODELS, PEAK_PARAMS, BKG_MODELS
 
 ATTRACTORS_PARAMS = {'distance': 20, 'prominence': None,
                      'width': None, 'height': None, 'threshold': None}
-FIT_PARAMS = {'method': 'leastsq', 'fit_negative': False, 'max_ite': 200}
+FIT_PARAMS = {'method': 'leastsq', 'fit_negative': False, 'max_ite': 200,
+              'coef_noise': 2}
 
 
 def create_model(model, model_name, prefix=None):
@@ -99,9 +101,14 @@ class Spectrum:
             Maximum number of iteration associated to the fitting process.
             An iteration consists in evaluating all the 'free' parameters once.
             Default is 200.
+        * coef_noise: float
+            Coefficient applied to the estimated noise amplitude to define a
+            threshold below which the fit weights are set to 0, and local
+            peak models are disabled .
+            Default is 2.
     result_fit: lmfit.ModelResult
         Object resulting from lmfit fitting. Default value is a 'None' object
-        (finction) that enables to address a 'result_fit.success' status.
+        (function) that enables to address a 'result_fit.success' status.
     """
 
     def __init__(self):
@@ -151,10 +158,6 @@ class Spectrum:
         for key in vars(self).keys():
             if key in keys:
                 setattr(self, key, model_dict[key])
-
-        # for key in ['fit_method', 'fit_negative', 'max_ite']:
-        #     if key in fit_kwargs:
-        #         setattr(self, key, fit_kwargs[key])
 
         if 'peak_models' in keys:
             self.peak_index = itertools.count(start=1)
@@ -323,13 +326,12 @@ class Spectrum:
         return peak_model
 
     def reassign_params(self):
-        """ Reassign fitted 'params' to the 'models' """
+        """ Reassign fitted 'params' values to the 'models' """
         for peak_model in self.peak_models:
             for key in peak_model.param_names:
                 param = self.result_fit.params[key]
                 name = key[4:]  # remove prefix 'mXX_'
-                peak_model.set_param_hint(name, value=param.value,
-                                          min=param.min, max=param.max)
+                peak_model.set_param_hint(name, value=param.value)
         if self.bkg_model is not None:
             for key in self.bkg_model.param_names:
                 param = self.result_fit.params[key]
@@ -431,7 +433,7 @@ class Spectrum:
         self.bkg_model.name2 = bkg_name
 
     def fit(self, fit_method=None, fit_negative=None, max_ite=None,
-            reinit_guess=True, **kwargs):
+            reinit_guess=True, coef_noise=None, **kwargs):
         """ Fit the Spectrum models """
         # update class attributes
         if fit_method is not None:
@@ -440,11 +442,17 @@ class Spectrum:
             self.fit_params['fit_negative'] = fit_negative
         if max_ite is not None:
             self.fit_params['max_ite'] = max_ite
+        if coef_noise is not None:
+            self.fit_params['coef_noise'] = coef_noise
 
         x, y = self.x, self.y
         weights = np.ones_like(x)
         if not self.fit_params['fit_negative']:
             weights[y < 0] = 0
+
+        ampli_noise = np.median(np.abs(y[:-1] - y[1:]) / 2)
[…]
