### 2024-02-05 215e677
manage default parameters for fwhm, fwhm_l, fwhm_r
