### 2024-02-05 a5f9484
revisit default parameters for fwhm, fwhm_l, fwhm_r
