### 2024-02-05 33d73c7
revisit noise_level estimation and application
