### 2024-02-05 4eefb7b
introduce 'xtol' as fitting parameters
