+++ fitspy/app/callbacks.py
@@ -659,6 +659,7 @@ class Callbacks:
         fit_params['coef_noise'] = params['coef_noise'].get()
         fit_params['max_ite'] = params['maximum_iterations'].get()
         fit_params['fit_method'] = params['fit_method'].get()
+        fit_params['xtol'] = params['xtol'].get()
 
         model_dict = self.current_spectrum.save()
         self.apply_model(model_dict=model_dict, fnames=fnames, fit_only=True,
