### 2024-02-05 aa3aa8f
redefine fwhm 'max' bounds to 200
