### 2024-02-05 597b54e
add ex_gui_peak_models_in_noise.py and related test
