+++ tests/test_gui_apply_model_to_all_new_spectra.py
@@ -13,9 +13,9 @@ def test_gui_apply_model_to_all(tmp_path):
     results = extract_results(dirname_res=tmp_path)
     # print(results)
 
-    refs = [[342.9457211120893, 378.09030982705525, 2.6380066104951028],
-            [342.9150412043217, 375.82154387665815, 2.745553299964265],
-            [342.83109830632253, 345.13789148228784, 2.7035947899485047]]
+    refs = [[342.9457600721084, 378.1308709613353, 2.63705383055729],
+            [342.915947718005, 376.40856906800724, 2.730686395444859],
+            [342.8311792192351, 345.1669426213287, 2.702655359738426]]
 
     for result, reference in zip(results, refs):
         assert result == approx(reference, rel=1e-3)
