+++ tests/test_nogui_auto_decomposition.py
@@ -14,9 +14,9 @@ def test_auto_decomposition():
     #     refs.append([np.sum(spectra[i].y), ampli])
     # print(refs)
 
-    refs = [[896641.9940061372, 28146.72777729365],
-            [890727.8824949837, 18920.21153480585],
-            [737968.1326228122, 9630.572508960488]]
+    refs = [[896641.9940061372, 28142.739608259562],
+            [890727.8824949837, 18900.368843277593],
+            [737968.1326228122, 9618.511108857907]]
 
     assert len(spectra) > 0
     for spectrum, ref in zip(spectra, refs):
