+++ tests/test_gui_users_defined_models.py
@@ -5,11 +5,11 @@ from examples.ex_gui_users_defined_models import ex_gui_users_models_from_txt
 from examples.ex_gui_users_defined_models import ex_gui_users_models_from_py
 from utils import extract_results, display_is_ok
 
-REFS = [[520.1862765603815, 915.2649558822233, 9.57229876472696],
-        [520.0548336803745, 895.8558065545101, 9.941802903292885],
-        [519.9924492642288, 892.7863717814992, 9.84735702289229],
-        [520.0449179477414, 903.3890860484772, 9.642083813467073],
-        [520.0348659274093, 931.4012074028958, 9.52151902889371]]
+REFS = [[520.1876881673046, 914.6124158904634, 9.567844475317354],
+        [520.0556606848833, 894.5345845967921, 9.952711474248666],
+        [519.9940849383077, 891.9396250396445, 9.841426352805783],
+        [520.0467482325055, 901.1376908860237, 9.648304792926787],
+        [520.0361487475782, 930.5679755394866, 9.522399137846483]]
 
 
 @pytest.mark.skipif(not display_is_ok(), reason="DISPLAY problem")
