+++ tests/test_gui_auto_decomposition.py
@@ -12,9 +12,9 @@ def test_gui_auto_decomposition(tmp_path):
     results = extract_results(dirname_res=tmp_path)
     # print(results)
 
-    refs = [[142.20616011719073, 28080.063179759956, 8.29468693999813],
-            [137.76510890707695, 18898.387213281483, 14.653606449091361],
-            [365.93083875575996, 10720.161126347022, 72.09201006313779]]
+    refs = [[142.20639528801823, 28071.78403149393, 8.301985144865942],
+            [137.7624794685612, 18872.059509383627, 14.717825498291326],
+            [365.9346475383543, 10717.74486424568, 72.09944175956193]]
 
     for result, reference in zip(results, refs):
         assert result == approx(reference, rel=2e-2)
