### 2024-02-05 #5868 @patquem
Thanks @phibeck for your supervision. From my side, I continue to improve Fitspy with a version v2024.3 that should be released soon. In the meantime, here's a new version of the draft that emphasizes (in ~one additional sentence) the ability to input 'user' models (available from v2024.2).
