### 2024-02-05 CEA-MetroCarac/fitspy#10 @maurov
@patquem thanks for your detailed reply and the improved documentation. It is fine for me now.
