### 2024-02-05 #5868 @maurov
@phibeck thanks for the reminder and I apologize for the long silence (busy period!). I have checked the latest improvements made by @patquem and I appreciate the efforts done. The documentation is more complete now and the paper was improved. Green light from my side to accept this contribution.
