### 2024-02-06 d9e2ab3
manage the progressbar independently of the backend
