### 2024-02-06 3e31c4e
bugfix: self.pbar_index instead of self.spectra.pbar_index
