### 2024-02-06 d63df39
remove checking on ncpus and len(fnames)
