### 2024-02-06 #5868 @maurov
@phibeck I did not have the time to check the performance yet. Nevertheless, I believe it fits the author's claim. @patquem if you could add a test file for checking performance, it could be useful.
