### 2024-02-07 e0a8b12
add outliers calculation, plotting and weighting in fit()
