+++ fitspy/spectra.py
@@ -7,6 +7,7 @@ import sys
 import time
 from threading import Thread
 from multiprocessing import Queue
+import numpy as np
 import matplotlib.pyplot as plt
 
 from fitspy.spectrum import Spectrum
@@ -35,6 +36,8 @@ class Spectra(list):
             super().__init__(spectra_list)
 
         self.spectra_maps = []
+        self.outliers_coef = 1.5
+        self.outliers_limit = None
         self.pbar_index = 0
 
     @property
@@ -51,6 +54,13 @@ class Spectra(list):
             spectra_all.extend(spectra_map)
         return spectra_all
 
+    def intensity(self):
+        """ Return the intensity array related to spectrum in spectra (ONLY) """
+        intensity = []
+        for spectrum in self:
+            intensity.append(spectrum.y)
+        return np.asarray(intensity)
+
     def get_objects(self, fname):
         """ Return spectrum and parent (spectra or spectra map)
             related to 'fname' """
@@ -66,6 +76,33 @@ class Spectra(list):
 
         print(f"{fname} not found in spectra")
         return None, None
+
+    def outliers_calculation(self, coef=1.5):
+        """ Outliers calculation from 'coef' * intensity_ref """
+
+        def outliers(intensity, spectra, coef):
+            shape = intensity.shape
+            inds = np.argsort(intensity, axis=0)[-5, :], np.arange(shape[1])
+            outliers_limit = coef * intensity[inds[0], inds[1]]
+            for spectrum in spectra:
+                outliers = np.where(spectrum.y > outliers_limit)[0]
+                spectrum.outliers = outliers.tolist()
+            return outliers_limit
+
+        # def outliers(intensity, spectra, coef):
+        #     mean = np.nanmean(intensity, axis=0)
+        #     std = np.nanstd(intensity, axis=0)
+        #     for spectrum in spectra:
+        #         z_score = (spectrum.y - mean) / std
+        #         spectrum.outliers = np.where(z_score > threshold)[0].tolist()
+        #     outliers_limit = threshold * std + mean
+        #     return outliers_limit
+
+        for spectra_map in self.spectra_maps:
+            spectra_map.outliers_limit = outliers(spectra_map.intensity,
+                                                  spectra_map, coef)
+        if len(self) > 0:
+            self.outliers_limit = outliers(self.intensity, self, coef)
 
     def save_results(self, dirname_res, fnames=None):
         """
