### 2024-02-07 07f5c0e
cleaning: remove old outliers() method based on z_score
