### 2024-02-08 1bba45b
revisit outliers management
