+++ fitspy/app/callbacks.py
@@ -83,8 +83,9 @@ class Callbacks:
         self.canvas.toolbar.update()
 
     def outliers_calculation(self):
-        """ Calculate the outliers"""
-        self.spectra.outliers_calculation(coef=self.outliers_coef.get())
+        """ Calculate the outliers (limit) """
+        coef = float(self.outliers_coef.get())
+        self.spectra.outliers_limit_calculation(coef=coef)
         if self.is_show_all:
             self.show_all()
         else:
@@ -111,18 +112,15 @@ class Callbacks:
 
         self.lines = []
         for spectrum in self.spectra.all:
-            x, y, inds = spectrum.x0, spectrum.y0, spectrum.outliers
-            self.lines.append(self.ax.plot(x, y, 'k-', lw=0.2, zorder=0)[0])
-            if inds is not None:
-                self.ax.plot(x[inds], y[inds], 'o', c='lime')
+            x0, y0 = spectrum.x0, spectrum.y0
+            self.lines.append(self.ax.plot(x0, y0, 'k-', lw=0.2, zorder=0)[0])
+            inds = np.where(y0 > spectrum.outliers_limit)[0]
+            if len(inds) > 0:
+                self.ax.plot(x0[inds], y0[inds], 'o', c='lime')
 
-        #  outliers
-        if self.figure_settings.params['plot_outliers_limit'].get() == 'On':
-            if self.spectra.outliers_limit is not None:
-                self.ax.plot(x, self.spectra.outliers_limit, 'r-', lw=2)
-            for spectra_map in self.spectra.spectra_maps:
-                if spectra_map.outliers_limit is not None:
-                    self.ax.plot(x, spectra_map.outliers_limit, 'r-', lw=2)
+        if self.figure_settings.params['plot_outliers_limit'].get() == 'On' \
+                and self.current_spectrum.outliers_limit is not None:
+            self.ax.plot(x0, self.current_spectrum.outliers_limit, 'r-', lw=2)
 
         self.canvas.draw()
 
@@ -362,6 +360,7 @@ class Callbacks:
         if fig_settings['plot_fit'].get() == 'On':
             show_attractors = self.attractors.get()
             show_outliers = fig_settings['plot_outliers'].get() == 'On'
+            show_outl_limit = fig_settings['plot_outliers_limit'].get() == 'On'
             show_neg_values = fig_settings['plot_negative_values'].get() == 'On'
             show_noise_level = fig_settings['plot_noise_level'].get() == 'On'
             show_baseline = fig_settings['plot_baseline'].get() == 'On'
@@ -369,6 +368,7 @@ class Callbacks:
             self.lines = spectrum.plot(self.ax,
                                        show_attractors=show_attractors,
                                        show_outliers=show_outliers,
+                                       show_outliers_limit=show_outl_limit,
                                        show_negative_values=show_neg_values,
                                        show_noise_level=show_noise_level,
                                        show_baseline=show_baseline,
