### 2024-02-08 b218a17
manage outliers in fit()
