### 2024-02-08 3d93cbe
add 'y_no_outliers' property used in several methods
