### 2024-02-08 53b906b
fixbug: incorrect names for fit methods
