### 2024-02-08 c4a64b7
add 'xtol' in ompatibility with 'old' key names
