### 2024-02-08 6357d94
limit the dill.dumps to save CPU costs
