### 2024-02-08 0e07595
move consecutive tasks to apply_model() in the same thread
