### 2024-02-08 c47bdce
move global variables in init.py
