### 2024-02-08 5bbce21
pass x0, y0 and outliers_limit to workers
