### 2024-02-08 25f64e0
save json with lists written on a single line
