### 2024-02-09 80d2b02
manage dict_from_tk_variables in the app
