### 2024-02-09 646c728
initialize attractors and fit settings from global variables defined in fitspy.py
