### 2024-02-09 f75b5a4
update model.json and rename spectrum.txt files
