+++ examples/ex_gui_2d_maps.py
@@ -20,10 +20,11 @@ def gui_2d_maps(dirname_res=None):
     unstr_map = os.path.join(DATA, '2D_maps', 'unordered_map.txt')
     appli.add_items(fnames=[str_map, unstr_map])
 
+    # appli.outliers_calculation() # spectra from the maps differ: DO NOT APPLY
+
     # automatic evaluation on the first 5 spectra
-    fnames = appli.spectra.fnames[:5]
-    appli.auto_eval(fnames=fnames)
-    # appli.auto_eval_all() # 4220 spectra to handle (could be long !!!)
+    appli.auto_eval(fnames=appli.spectra.fnames[:5])
+    # appli.auto_eval_all() # 1520+2700 = 4220 spectra to handle (could be long)
 
     # save and destroy for pytest
     if dirname_res is not None:
