+++ examples/data/spectra_2/InP-2_75B-P03.txt
@@ -1,575 +1,0 @@
-958.793	155.616
-957.299	224.221
-955.805	172.334
-954.311	157.262
-952.816	175.649
-951.32	167.27
-949.826	187.325
-948.332	185.635
-946.836	193.978
-945.342	173.895
-943.846	175.55
-942.352	170.518
-940.855	172.172
-939.359	183.855
-937.863	177.151
-936.367	162.093
-934.871	187.14
-933.375	218.864
-931.879	177.077
-930.381	167.036
-928.885	155.327
-927.389	178.69
-925.891	180.34
-924.393	170.302
-922.896	180.299
-921.398	143.555
-919.9	178.589
-918.402	176.9
-916.904	158.524
-915.406	128.472
-913.906	170.164
-912.408	180.152
-910.91	170.123
-909.41	190.115
-907.91	160.077
-906.412	185.066
-904.912	143.366
-903.412	176.685
-901.912	166.663
-900.412	159.976
-898.912	169.953
-897.41	173.263
-895.91	166.577
-894.408	168.221
-892.908	178.191
-891.406	159.851
-889.904	169.819
-888.402	183.113
-886.9	173.101
-885.398	169.749
-883.896	179.71
-882.395	174.694
-880.891	161.361
-879.389	164.665
-877.885	187.925
-876.381	177.921
-874.877	196.184
-873.373	209.455
-871.869	156.237
-870.365	179.481
-868.861	196.071
-867.355	161.153
-865.852	166.113
-864.346	172.732
-862.84	161.081
-861.334	177.661
-859.828	156.053
-858.322	159.349
-856.816	184.219
-855.309	177.553
-853.803	177.526
-852.295	174.181
-850.789	179.13
-849.281	157.543
-847.773	180.732
-846.266	182.361
-844.756	174.045
-843.248	193.905
-841.74	193.874
-840.23	150.767
-838.721	165.651
-837.211	177.218
-835.703	173.878
-834.191	188.751
-832.682	200.308
-831.172	175.448
-829.66	182.038
-828.15	186.972
-826.639	175.361
-825.127	190.218
-823.615	186.879
-822.104	162.045
-820.592	148.792
-819.08	186.785
-817.566	198.322
-816.053	165.24
-814.541	185.038
-813.027	185.006
-811.514	170.11
-810	193.199
-808.484	156.844
-806.971	166.721
-805.455	199.7
-803.941	178.214
-802.426	178.183
-800.91	168.254
-799.395	189.665
-797.877	171.493
-796.361	202.787
-794.846	179.674
-793.328	184.586
-791.811	181.258
-790.293	191.111
-788.775	202.607
-787.258	215.746
-785.74	189.361
-784.221	187.68
-782.701	200.815
-781.184	177.738
-779.664	162.897
-778.145	187.544
-776.625	199.024
-775.104	179.253
-773.584	198.951
-772.063	189.051
-770.541	164.361
-769.02	165.974
-767.498	188.946
-765.977	167.556
-764.455	190.518
-762.932	172.419
-761.41	165.82
-759.887	164.148
-758.363	173.964
-756.84	183.776
-755.316	167.336
-753.791	190.268
-752.268	157.434
-750.742	163.962
-749.217	162.292
-747.691	168.817
-746.166	185.172
-744.641	185.136
-743.113	188.377
-741.588	186.704
-740.061	167.019
-738.533	157.164
-737.006	157.134
-735.479	178.378
-733.951	197.978
-732.422	171.766
-730.895	173.368
-729.365	199.498
-727.836	220.713
-726.307	194.517
-724.777	212.456
-723.246	176.468
-721.717	160.097
-720.186	171.499
[…]

[…]
