### 2024-02-09 d9b1e30
remove dict_has_tk_variable
