### 2024-02-09 3877502
change ref in test_gui_auto_decomposition.py
