+++ tests/test_gui_auto_decomposition.py
@@ -12,9 +12,9 @@ def test_gui_auto_decomposition(tmp_path):
     results = extract_results(dirname_res=tmp_path)
     # print(results)
 
-    refs = [[142.20639528801823, 28071.78403149393, 8.301985144865942],
-            [137.7624794685612, 18872.059509383627, 14.717825498291326],
-            [365.9346475383543, 10717.74486424568, 72.09944175956193]]
+    refs = [[142.2063952870293, 28071.7840261713, 8.30198514788677],
+            [137.76247944776298, 18872.059537827474, 14.717825224085768],
+            [365.9688879244104, 10719.133628864241, 71.84603593278625]]
 
     for result, reference in zip(results, refs):
         assert result == approx(reference, rel=2e-2)
