### 2024-02-09 c180e54
split tabview into paramsview and statsview
