### 2024-02-10 55e821e
bugfix: 'fit_method' replaces by 'method'
