### 2024-02-12 a7b54b1
revisit FileSelector with single input and add select_item()
