+++ fitspy/app/utils.py
@@ -5,7 +5,7 @@ import os
 import glob
 from itertools import groupby, count
 from tkinter import (LabelFrame, Frame, Label, Entry, Canvas, Scrollbar, Button,
-                     Listbox, Variable, StringVar, TclError)
+                     Listbox, TclError)
 from tkinter import W, E, HORIZONTAL, EXTENDED, BOTTOM, X, Y, LEFT, RIGHT, END
 from tkinter import filedialog as fd
 from tkinter.messagebox import showerror
@@ -199,214 +199,159 @@ class FilesSelector:
 
     Attributes
     ----------
-    ninputs: int
-        Number of inputs to consider
     lbox: list of Listbox object
-        List of 'ninputs'-Listbox associated to the selected files
+        Listbox associated to the selected files
     filenames: list of str
-        List of 'ninputs'-list of filenames to work with
-    files_extension: iterable of str
-        Extension used for the files selection
-    self.dirname_res: str
-        Directory pathname where to find results associated to the files
+        List of of filenames to work with
 
     Parameters
     ----------
     root: Tk.widget
         The main window associated to the FileSelector
-    ninputs: int, optional
-        Number of inputs to consider
-    files_extensions: str, optional
-        Extension used for files selection.
-    dirname_results: str, optional
-        Directory pathname where to find results associated to the files
-        'DEFAULT' value refers to './results' directory associated to the 'file'
-        location
     lbox_size: list of 2 ints, optional
         Size (width, height) of the Listbox 'lbox'. Default value is [30, 15]
-    options: dict, optional
-        Dictionary of options
-    tab_options: Frame object
-        Tab dedicated to options parameters settings
     """
 
-    def __init__(self, root,
-                 ninputs=1,
-                 files_extensions='*.txt',
-                 dirname_results='DEFAULT',
-                 lbox_size=None,
-                 options=None, tab_options=None):
+    def __init__(self, root, lbox_size=None):
 
         lbox_size = lbox_size or [30, 15]
-        if options is None:
-            options = {}
 
-        self.ninputs = ninputs
-        self.lbox = []
-        self.options = options
+        self.filenames = []
 
-        self.filenames = [[] for _ in range(ninputs)]
+        # create buttons and listbox
 
-        self.options['files_extension'] = files_extensions
-        self.options['dirname_results'] = dirname_results
+        Button(root, text="Select Files",
+               command=self.select_files). \
+            grid(column=0, row=0, padx=0, pady=5)
+        Button(root, text="Select Dir.",
+               command=self.select_dir). \
+            grid(column=1, row=0, padx=0, pady=5)
+        Button(root, text="Remove",
+               command=self.remove). \
+            grid(column=2, row=0, padx=0, pady=5)
+        Button(root, text="Remove all",
+               command=self.remove_all). \
+            grid(column=3, row=0, padx=0, pady=5)
 
-        # create buttons/listbox associated to each input
-        for i in range(self.ninputs):
-            inc = 3 * i
+        Button(root, text="▲", command=lambda: self.move('up')). \
+            grid(column=1, row=1, pady=0, sticky=E)
+        Button(root, text="▼", command=lambda: self.move('down')). \
+            grid(column=2, row=1, pady=0, sticky=W)
 
-            Button(root, text="Select Files",
-                   command=lambda ind=i: self.select_files(ind)). \
-                grid(column=0, row=0 + inc, padx=0, pady=5)
-            Button(root, text="Select Dir.",
-                   command=lambda ind=i: self.select_dir(ind)). \
[…]
