+++ fitspy/app/callbacks.py
@@ -124,7 +124,7 @@ class Callbacks:
         def on_press(event):
             """ Highlight spectra close to the mouse click position """
             if self.lines is not None and event.inaxes == self.ax:
-                fselector.lbox[0].selection_clear(0, END)
+                fselector.lbox.selection_clear(0, END)
                 # reassign standard properties to the previous nearest lines
                 for line in self.nearest_lines:
                     line.set(linewidth=0.2, color='k', zorder=0)
@@ -132,21 +132,21 @@ class Callbacks:
                 self.nearest_lines, labels = [], []
                 for i, line in enumerate(self.lines):
                     if line.contains(event)[0]:
-                        fselector.lbox[0].selection_set(i)
+                        fselector.select_item(i)
                         if len(self.nearest_lines) < 10:
                             color = CMAP(len(self.nearest_lines))
                             line.set(linewidth=1, color=color, zorder=1)
                             self.nearest_lines.append(line)
-                            label = os.path.basename(fselector.filenames[0][i])
+                            label = os.path.basename(fselector.filenames[i])
                             labels.append(label)
                         else:
                             title = "The nearest lines exceed 10.\n"
                             title += "  Show the first 10 ones"
                         # adapt the scrollbar cursor position to the first line
                         if len(self.nearest_lines) == 1:
-                            cursor_position = i / len(fselector.filenames[0])
-                            fselector.lbox[0].yview_moveto(cursor_position)
-                            fselector.lbox[0].update()
+                            cursor_position = i / len(fselector.filenames)
+                            fselector.lbox.yview_moveto(cursor_position)
+                            fselector.lbox.update()
                 self.ax.legend(self.nearest_lines, labels, title=title, loc=1)
                 self.canvas.draw()
 
@@ -176,8 +176,7 @@ class Callbacks:
                 os.makedirs(dirname_res)
 
         if os.path.isdir(dirname_res):
-            self.spectra.save_results(dirname_res,
-                                      self.fileselector.filenames[0])
+            self.spectra.save_results(dirname_res, self.fileselector.filenames)
 
     def save_figures(self, dirname_fig=None):
         """ Save all spectra figures in .png files """
@@ -190,7 +189,7 @@ class Callbacks:
         if os.path.isdir(dirname_fig):
             bounds = (self.ax.get_xlim(), self.ax.get_ylim())
             self.spectra.save_figures(dirname_fig,
-                                      fnames=self.fileselector.filenames[0],
+                                      fnames=self.fileselector.filenames,
                                       bounds=bounds)
 
     def load_model(self, fname_json=None):
@@ -223,12 +222,11 @@ class Callbacks:
             return
 
         if fnames is None:
-            fnames = self.fileselector.filenames[0]
+            fnames = self.fileselector.filenames
 
         if selection:
-            fnames = self.fileselector.filenames[0]
-            fnames = [fnames[i]
-                      for i in self.fileselector.lbox[0].curselection()]
+            fnames = self.fileselector.filenames
+            fnames = [fnames[i] for i in self.fileselector.lbox.curselection()]
 
         nfiles = len(fnames)
         ncpus = self.get_ncpus(nfiles=nfiles)
@@ -279,16 +277,16 @@ class Callbacks:
 
     def save_selection(self, fname_json=None):
         """ Save selected spectra models in a .json file """
-        inds = self.fileselector.lbox[0].curselection()
+        inds = self.fileselector.lbox.curselection()
[…]
