+++ tests/utils.py
@@ -13,7 +13,7 @@ from fitspy.utils import hsorted
 def extract_results(dirname_res):
     """ Return results extracted from fit parameters .csv files """
     fnames = glob.glob(os.path.join(dirname_res, "*.csv"))
-    fnames = hsorted([x for x in fnames if "stat" not in x])
+    fnames = hsorted([x for x in fnames if "results.csv" not in x])
     results = []
     for fname in fnames:
         dfr = pd.read_csv(fname, sep=';', header=1)
