### 2024-02-12 18e7ae1
manage setattr() for dictionaries
