### 2024-02-13 3d9fcf4
bugfix in manage setattr() for dictionaries
