+++ fitspy/spectrum.py
@@ -164,12 +164,12 @@ class Spectrum:
             self.fit_params['xtol'] = model_dict.pop('xtol')
 
         for key, val in vars(self).items():
-            if isinstance(val, dict):
-                for key2 in val.keys():
-                    if key2 in model_dict[key].keys():
-                        val[key2] = model_dict[key][key2]
-            else:
-                if key in keys:
+            if key in keys:
+                if isinstance(val, dict) and key:
+                    for key2 in val.keys():
+                        if key2 in model_dict[key].keys():
+                            val[key2] = model_dict[key][key2]
+                else:
                     setattr(self, key, model_dict[key])
 
         if 'peak_models' in keys:
