### 2024-02-14 a353369
add 'force_terminal_exit' attribute
