### 2024-02-16 #5868 @FCMeng
@phibeck @patquem
Thanks for the reminder. My review process is almost done. I agree this is an excellent work and fits the scope for JOSS. I do have a few questions when testing with my own data.

signal.find_peaks() is used to find the peaks in the spectra. If the resolution is not good enough, some peaks might be ignored in this process, especially small "shoulder" peak near the main peak. In this case, is it possible to do an auto interpolation first and then fit on the interpolated spectra? Or is it possible to give the users option to interpolate the spectra in fitspy?

I noticed there is small portion of todo list on the developer part in the main documentation. A coding/debugging guideline for developers/contributors is strongly recommended, which benefit the community.

The supported operating systems and python version should be provided. When I was trying to use fitspy on Macbook with M2 chip, the installation process succeeded but have issue when launching fitspy.

Is the size of spectra figure in the GUI fixed? I am facing an issue use the GUI on a Linux machine (Rocky OS),  where the size of the spectra figure will not be reshaped automatically when I resize the GUI window. In this case, I can not see the whole spectra or some part of the menu on the right side of the spectra figure if the GUI window is too small.
