### 2024-02-21 36a5ab3
Rename 'peaks' frame into 'Fitting'
