### 2024-02-21 5617ffa
add doc for fit()
