+++ doc/user_guide/outputs.rst
@@ -1,0 +1,156 @@
+Outputs
+=======
+
+Fitspy provides two types of outputs:
+
+* those related to the **models**,
+
+* and those related to the fits (**parameters** of the fitted models and **statistics**).
+
+
+Outputs models
+--------------
+
+The **output models** consist of .json files in which all parameters of the spectrum and spectra objects have been saved so as to be able:
+
+* either to recreate the previously obtained results **identically**
+
+* or to apply the first spectrum model to other data sets (without considering the :code:`fname` value)::
+
+    {
+       "0": {
+          "fname": "../examples/data/spectra_2/spectrum_2_1.txt",
+          "range_min": 200.0,
+          "range_max": 957.299,
+          "norm_mode": "Attractor",
+          "norm_position_ref": 685,
+          "peaks": [88, 64, 307, 188, 210, 262, 113, 168, 285, 145, 31, 231, 330, 464, 431, 394, 366, 11],
+          "peaks_params": {
+             "distance": 20,
+             "prominence": 0,
+             "width": 0,
+             "height": 0,
+             "threshold": 0
+          },
+          "models_labels": ["", "Peak IV", "Peak II", "", "", "Peak I", "Peak III", "", ""],
+          "bkg_model": null,
+          "fit_method": "Leastsq",
+          "fit_negative": true,
+          "max_ite": 200,
+          "attached": true,
+          "baseline": {
+             "points": [[201.1400035483871, 954.8547838709676], [164.91378720565154, 228.197032967033]],
+             "mode": "Linear",
+             "order_max": 2,
+             "distance": 100,
+             "sigma": 0,
+             "attached": true,
+             "is_subtracted": true
+          },
+          "models": {
+             "0": {
+                "Lorentzian": {
+                   "ampli": {
+                      "value": 377.8408958125738,
+                      "min": 0,
+                      "max": Infinity,
+                      "vary": 1,
+                      "expr": null
+                   },
+                   "x0": {
+                      "value": 342.94548419455225,
+                      "min": 322.195,
+                      "max": 362.195,
+                      "vary": 1,
+                      "expr": null
+                   },
+                   "fwhm": {
+                      "value": 2.643513972868272,
+                      "min": 0,
+                      "max": 200,
+                      "vary": 1,
+                      "expr": null
+                   }
+                }
+             },
+             "1": {
+                "Lorentzian": {
+                   "ampli": {
+                      "value": 98.30555093353449,
+                      "min": 0,
+                      "max": Infinity,
+                      "vary": 1,
+                      "expr": null
+                   },
+                   ...
+             },
+             ...
+
+Fit results
+-----------
+
+Once fitted, a spectra data set can generate several files related to the fitting statistics and the fitted parameters.
+
+In a 'results' directory (to be specified by the user), the **statistics** files are built when clicking on :code:`Save Results` in the GUI, according to the following model: {spectrum basename}**_stats.txt** file).
+
+The same for the **fitted parameters** attached to each spectrum: {spectrum basename}**.csv** file)
+
+* Example of **statistics** returned in *ordered_map.txt  X=1757.0  Y=65_stats.txt* (obtained from *examples/ex_gui_users_defined_models_2d_map.py*)::
+
[…]
