### 2024-02-22 c2909e1
fix incorrect outliers_limit x-range in plot()
