+++ doc/dev_guide/dev_notes.rst
@@ -1,16 +1,36 @@
 Developers Notes
 ================
 
-Developers are warmly encouraged to collaborate and contribute to the ongoing improvement of Fitspy.
+Fitspy is a new project (started in 2023).
+
+Developers are warmly encouraged to collaborate and contribute to the ongoing improvement of Fitspy. Contributions are welcome in the form of bug reports, documentation, code , feature requests, and more...
+
+... keeping in mind that Fitspy must maintain its primary goal: to remain as simple, generic and user-friendly as possible :)
+
+
+Issues
+------
+
+The `issue tracker <https://github.com/CEA-MetroCarac/fitspy/issues>`_ can be used to report bugs or propose new features. When reporting a bug, the following is useful:
+
+* give a minimal example demonstrating the bug,
+
+* copy and paste the error traceback.
+
+Pull Requests
+-------------
+
+If you want to contribute to improve the Fitspy source code, you can send us a `pull request <https://github.com/CEA-MetroCarac/fitspy/pulls>`_ against the main branch. Small bug fixes and corrections to the user guide are typically a good starting point. But don’t hesitate also for significant code contributions, such as support for a new file format - if needed, we’ll help you to get the code ready to common standards.
+
 
 TODO List
 ---------
 
 * investigate erratic timelaps when opening files with tkinter.filedialog
-* improve convergence speed during spectra dataset fitting
 * update dynamically the 2D map figure during fitting
-* add option to remove overshoots in profiles ('cosmic rays')
-* fix expected floating-point in sigma with interactive_entry
 * make widgets size depending on the users screen size
 * enable widget stretching
-* change Tkinter for pyQt or PySide :)
+* **change Tkinter for pyQt or PySide** :) (task scheduled for mid-2024)
+* create a fitspy icon for the taskbar
+* create an 'About'
+* add the progress bar inside the GUI
