### 2024-02-23 1c93561
update logo in .rst
