### 2024-02-23 81b36a9
update first figure in .rst
