+++ README.rst
@@ -19,8 +19,7 @@
 
    <br>
 
-**Fitspy** is a generic tool dedicated to **fit** **sp**ectra in **py**thon
-with a GUI that aims to be as simple and intuitive to use as possible.
+**Fitspy** is a generic tool dedicated to **fit sp**ectra in **py**thon with a GUI that aims to be as simple and intuitive to use as possible.
 
 
 .. figure::  https://cea-metrocarac.github.io/fitspy/fitspy.png
@@ -33,72 +32,81 @@ with a GUI that aims to be as simple and intuitive to use as possible.
 
 Processed spectra may be independent of each other or may result from 2D-maps
 acquisitions.
-<p align="center" width="100%">
-    <img align="center" width="30%" src=https://cea-metrocarac.github.io/fitspy/2d-map.png> <br>
-    <em>Example of fitspy 2D-map frame interacting with the main GUI.</em> 
 
-</p>
+.. figure::  https://cea-metrocarac.github.io/fitspy/2d-map.png>
+   :align:   center
+   :width:   75%
 
-The predefined peak models considered in Fitspy are  `Gaussian`, `Lorentzian`, `Asymetric Gaussian`, `Asymetric Lorentzian` and `Pseudovoigt`.
+   Example of fitspy 2D-map frame interacting with the main GUI.
 
-A `constant`, `linear`, `parabolic` or `exponential` background model can also be added in the fitting.
+.. raw:: html
 
-In both cases, `user-defined models` can be added.
+   <br>
+
+The predefined peak models considered in Fitspy are  :code:`Gaussian`, :code:`Lorentzian`, :code:`Asymetric Gaussian`, :code:`Asymetric Lorentzian` and :code:`Pseudovoigt`.
+
+A :code:`constant`, :code:`linear`, :code:`parabolic` or :code:`exponential` background model can also be added in the fitting.
+
+In both cases, :code:`user-defined models` can be added.
 
 Fitspy main features:
 
-- Fitspy uses the [lmfit](https://github.com/lmfit/lmfit-py) library to fit the spectra
+- Fitspy uses the `lmfit <https://github.com/lmfit/lmfit-py>`_ library to fit the spectra
 - The fit processing can be multi-threaded
 - Bounds and constraints can be set on each peaks models parameter.
 - From an automatic noise level estimation, according to the local noise, peak models can be automatically deactivated.
 - Fitspy also includes automatic outlier detection to be excluded during the fitting process.
 
-All actions allowed with the GUI can be executed in script mode (see examples [here](https://github.com/CEA-MetroCarac/fitspy/tree/main/examples)).
+All actions allowed with the GUI can be executed in script mode (see examples `here <https://github.com/CEA-MetroCarac/fitspy/tree/main/examples>`_).
 These actions (like baseline and peaks definition, parameters constraints, ...) can be saved in a `Fitspy model` and replayed as-is or applied to other new spectra datasets.
 
 
-### Installation
+Installation
+------------
 
-```
-pip install fitspy
-```
+    pip install fitspy
 
-### Tests and examples execution
 
-```
-pip install pytest
-git clone https://github.com/CEA-MetroCarac/fitspy.git
-cd fitspy
-pytest
-python example/ex_gui_auto_decomposition.py
-...
-```
+Tests and examples execution
+----------------------------
 
-### Quick start
+
+    pip install pytest
+    git clone https://github.com/CEA-MetroCarac/fitspy.git
+    cd fitspy
+    pytest
+    python example/ex_gui_auto_decomposition.py
+    ...
+
+
+Quick start
+-----------
 
 Launch the application:
 
-```
-fitspy
-```
+::
+
+    fitspy
+
 
[…]
