### 2024-02-23 dc878db
fix bug when outliers_limit is None
