### 2024-02-26 21f3f65
fix bug from frame 'Peaks' renaming into 'Fitting'
