### 2024-02-26 9a09936
change readme = README.md to README.rst
