### 2024-02-27 e76b019
add auto_mode.rst / update doc
