+++ doc/index.rst
@@ -10,6 +10,7 @@ Welcome to Fitspy's documentation!
    user_guide/peak_models.rst
    user_guide/fitting.rst
    user_guide/workflow.rst
+   user_guide/auto_mode.rst
    user_guide/gui.rst
    user_guide/outputs.rst
 
