+++ doc/release_notes/release_notes.rst
@@ -2,7 +2,7 @@ Release Notes
 =============
 
 
-Version 2024.3 (February 23, 2024)
+Version 2024.4 (February 27, 2024)
 ----------------------------------
 
 New features:
