+++ README.md
@@ -1,0 +1,101 @@
+[![PyPI](https://img.shields.io/pypi/v/fitspy?label=pypi%20package)](
+https://pypi.org/project/fitspy/)
+[![Github](https://img.shields.io/badge/GitHub-GPL--3.0-informational)](
+https://github.dev/CEA-MetroCarac/fitspy)
+[![Doc](https://img.shields.io/badge/%F0%9F%95%AE-docs-green.svg)](
+https://cea-metrocarac.github.io/fitspy/doc/index.html)
+
+
+<p align="center" width="100%">
+    <img align="center" width=250 src=https://cea-metrocarac.github.io/fitspy/logo.png>
+</p>
+
+
+**Fitspy** is a generic tool dedicated to **fit sp**\ ectra in **py**\ thon with a GUI that aims to be as simple and intuitive to use as possible.
+
+<p align="center" width="100%">
+    <img align="center" width="75%" src=https://cea-metrocarac.github.io/fitspy/fitspy.png>
+</p>
+
+Processed spectra may be independent of each other or may result from 2D-maps
+acquisitions.
+
+<p align="center" width="100%">
+    <img align="center" width="30%" src=https://cea-metrocarac.github.io/fitspy/2d-map.png> <br>
+    <em>Example of fitspy 2D-map frame interacting with the main GUI.</em> 
+</p>
+
+The predefined peak models considered in Fitspy are  `Gaussian`, `Lorentzian`, `Asymetric Gaussian`, `Asymetric Lorentzian` and `Pseudovoigt`.
+
+A `constant`, `linear`, `parabolic` or `exponential` background model can also be added in the fitting.
+
+In both cases, `user-defined models` can be added.
+
+Fitspy main features:
+
+- Fitspy uses the [lmfit](https://github.com/lmfit/lmfit-py) library to fit the spectra
+- The fit processing can be multi-threaded
+- Bounds and constraints can be set on each peaks models parameter
+- From an automatic noise level estimation, according to the local noise, peak models can be automatically deactivated
+- Fitspy also includes automatic outlier detection to be excluded during the fitting process
+
+All actions allowed with the GUI can be executed in script mode (see examples [here](https://github.com/CEA-MetroCarac/fitspy/tree/main/examples)).
+These actions (like baseline and peaks definition, parameters constraints, ...) can be saved in a `Fitspy model` and replayed as-is or applied to other new spectra datasets.
+
+
+## Installation
+
+```
+pip install fitspy
+```
+
+*(See the [documentation](https://cea-metrocarac.github.io/fitspy/doc/user_guide/intro.html#install-and-launching) for more details)*
+
+
+## Tests and examples execution
+
+```
+pip install pytest
+git clone https://github.com/CEA-MetroCarac/fitspy.git
+cd fitspy
+pytest
+python example/ex_gui_auto_decomposition.py
+```
+
+*(See the [documentation](https://cea-metrocarac.github.io/fitspy/doc/user_guide/intro.html#install-and-launching) for more details)*
+
+
+## Quick start
+
+Launch the application:
+
+```
+fitspy
+```
+
+Then, from the top to the bottom of the right panel:
+
+- `Select` file(s)
+- *(Optional)* Define the **X-range**
+- Define the baseline to `subtract` *(left or right click on the figure to add or delete (resp.) a baseline point)*
+- *(Optional)* Normalize the spectrum/spectra
+- Click on the `Fitting` panel to activate it
+- Select `Peak model` and add peaks *(left or right click on the figure to add or delete (resp.) a peak)*
+- *(Optional)* Add a background (**BKG model**) to be fitted
+- *(Optional)* Use **Parameters** to set bounds and constraints
[…]
