+++ pyproject.toml
@@ -8,7 +8,7 @@ name = "fitspy"
 dynamic = ["version"]
 authors = [{ name = "Patrick Quéméré", email = "patrick.quemere@cea.fr" }]
 description = "Fitspy: a generic tool to fit spectra in python"
-readme = "README.rst"
+readme = "README.md"
 requires-python = ">=3.7"
 keywords = ["Fitspy", "fit", "spectra", "spectrum", "map", "1D", "2D",
     "decomposition", "Gaussian", "Lorentzian", "Pseudovoigt", "GUI"]
