+++ doc/user_guide/intro.rst
@@ -79,7 +79,7 @@ Once the `Fitspy` GUI has opened, from top to bottom of the right panel, the str
 - (`1 <files_selection.html>`_) **Select file(s) from** :code:`Select Files`  **or**  :code:`Select Dir`
 - (`6 <fitting.html>`_) **Click on the** :code:`Fitting` **panel to activate it (if not)**
 - (`7 <fitting.html>`_) **Select a** :code:`Peak model`
-- (`8 <fitting.html>`_) **Select a peak point on the main figure** *(with left/right click on the figure to add/delete a baseline point)*
+- (`8 <fitting.html>`_) **Select a peak point on the main figure** *(with left/right click on the figure to add/delete a peak point)*
 - Repeat the last two actions as needed
 - (`11 <fitting.html>`_) :code:`Fit Selec.` **the spectrum/spectra selected in the files selector widget or** :code:`Fit All` **the spectra**
 - (`12 <fitting.html>`_) :code:`Save Results` (**fitted parameters** and **statistics**)
