### 2024-02-27 #5868 @FCMeng
@patquem
Thanks for your detailed responses.

    Let me rephrase the first question in this way: can signal.find_peaks() always return the correct peaks? If not, any ways to improve it?
    The last minor question I was referring here: https://github.com/CEA-MetroCarac/fitspy?tab=readme-ov-file#tests-and-examples-execution
    where the last command python example/ex_gui_auto_decomposition.py should be python examples/ex_gui_auto_decomposition.py.

@phibeck My bad! I forgot to switch account. I have deleted the previous comment and repost it here with the correct account.
