### 2024-02-28 #5868 @patquem
@FCMeng

`1. can signal.find_peaks() always return the correct peaks? If not, any ways to improve it?`
find_peaks() return peaks according to its possibilities (finding local maxima by simple comparison of neighboring values, given the different parameters that can be adjusted by the user). This method may work well for distinct and separated peaks. However, for merged peaks, the task becomes significantly more complex and may require users' expertise. (Practically, people using fitspy around me, set the peaks manually).

`2. exampleS in the README`
You are right. This has been fixed.

@maurov

`Performance (complement)`
As previously mentioned, Fitspy's intrinsic performance aligns with that of lmfit (along with multithreading capabilities). CPU time varies depending on factors such as the number of spectra, the number of points per spectrum, the number of peaks to fit ... and of course the machine performance including its multiprocessing capabilities.
For example, fitting all the 1520 spectra of ex_gui_users_defined_models_2d_map.py with the pre-defined model (5 peaks)  takes ~22 seconds with 32 CPUs in our Windows server.
