### 2024-03-13 #5868 @patquem
Hi @phibeck,

Regarding the last draft published the 05/02/24 with ID : [10.21105.joss.05868.pdf](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05868/joss.05868/10.21105.joss.05868.pdf) the authors and affiliations (including ORCIDs) are **OK** for me.
The reference Fitspy version for JOSS is the **2024.4**.
This release has been archived on Zenodo : [https://zenodo.org/doi/10.5281/zenodo.10812332](https://zenodo.org/doi/10.5281/zenodo.10812332) with title and author list matching those in the JOSS paper.
The license listed for the archive is the same as the software license.

Hoping it will be okay with you to finalize the article publishing :)
