### 2024-03-13 c30fd13
add zenodo DOI and acknowledgements
