### 2024-03-14 ce875a6
update calls to spectrum functions
