### 2024-03-14 ca9168c
add show_peak_models and show_result as activation keywords
