+++ doc/index.rst
@@ -1,5 +1,20 @@
-Welcome to Fitspy's documentation!
-====================================
+|docs|_  |github|_  |pypi|_  |joss|_
+
+.. |docs| image:: https://img.shields.io/badge/%F0%9F%95%AE-docs-green.svg
+.. _docs: https://cea-metrocarac.github.io/fitspy/doc/index.html
+
+.. |github| image:: https://img.shields.io/badge/GitHub-GPL--3.0-informational
+.. _github: https://github.dev/CEA-MetroCarac/fitspy
+
+.. |pypi| image:: https://img.shields.io/pypi/v/fitspy?label=pypi%20package
+.. _pypi: https://pypi.org/project/fitspy
+
+.. |joss| image:: https://joss.theoj.org/papers/971a02868d903c0b7c0cbc3d1cd3d139/status.svg
+.. _joss: https://joss.theoj.org/papers/971a02868d903c0b7c0cbc3d1cd3d139
+
+
+Welcome to the Fitspy documentation !
+=====================================
 
 .. toctree::
    :caption: User Guide
