+++ doc/index.rst
@@ -1,13 +1,16 @@
-|docs|_  |github|_  |pypi|_  |joss|_
+|docs|_  |github|_  |pypi|_  |doi|_  |joss|_
 
 .. |docs| image:: https://img.shields.io/badge/%F0%9F%95%AE-docs-green.svg
 .. _docs: https://cea-metrocarac.github.io/fitspy/doc/index.html
 
 .. |github| image:: https://img.shields.io/badge/GitHub-GPL--3.0-informational
-.. _github: https://github.dev/CEA-MetroCarac/fitspy
+.. _github: https://github.com/CEA-MetroCarac/fitspy
 
 .. |pypi| image:: https://img.shields.io/pypi/v/fitspy?label=pypi%20package
 .. _pypi: https://pypi.org/project/fitspy
+
+.. |doi| image:: https://zenodo.org/badge/DOI/10.5281/zenodo.10812333.svg
+.. _doi: https://doi.org/10.5281/zenodo.10812333
 
 .. |joss| image:: https://joss.theoj.org/papers/971a02868d903c0b7c0cbc3d1cd3d139/status.svg
 .. _joss: https://joss.theoj.org/papers/971a02868d903c0b7c0cbc3d1cd3d139
