### 2024-03-20 #5868 @phibeck
Hi @patquem. A couple of comments:
- the version is somehow tagged as `v2024.4` in the git repo link to the pypi package. Just want to confirm that the version is `2024.4`. Additionally, it may help to tag the same version in git.
- we do require the licenses to match between the archive and the software repo, but they are `GPLv3` for zenodo/pypi and `AGPL-3.0` for the repo. Please correct this and let me know
- reference `Meyer et al. 2023` doesn't have a DOI, and doesn't contain the image used in the manuscript. Is there a paper draft on the arxiv that can be used instead?
