### 2024-03-28 #5868 @patquem
Hello @phibeck,

Regarding the first and second points, I have standardized the version number between the PyPI repository and the badges (it is version 2024.4), as well as the license (it is GPLv3). On the last point, apart from the abstract referenced in the article, Thibault Meyer's work presented at the RamanFest2023 conference has not yet been published in a paper.
