### 2024-04-01 #5868 @phibeck
@patquem Thank you for correcting the version number and unifying the license! Here are a few more comments/suggestions for the manuscript. Please have a look when you have a moment.

- Throughout the text: I find it more intuitive to speak of a single "user" instead of "users", but this is a personal preference
- line 5: I would merge this sentence with the following paragraph, it looks unusual to me to have this separate
- line 7f.: I would add at the end of the sentence "as peak models"
- Fig. 1: It would be helpful to add a little more detail, i.e. that the noisy line is some experimental measurement and the solid lines of different colors are returned after running Fitspy
- line 14f.: I am not sure I would say "chemical" composition here, but rather just composition. There are certainly use cases for tools like Fitspy where peak positions aren't a matter of chemical composition, but other parameters.
- line 42: none of them "seem" (without "s")
- line 47: at the time of writing
- line 49: "two-column base format"
- line 49: I would rephrase slightly: "base format, with the first column containing the physical support..."
- line 52: "is associated with"
- line 53: "identified by its grid coordinates (X,Y)"
- line 55 and 56: "file selection widget"
- line 59f: please correct if this isn't what you mean: "...Attractors capability, which automatically returns the local maxima.."
- line 64: "...so that the spectra have a common, flat zero-intensity level"
- line 65: either "user, who positions" or "users, who position"
- line 66: "with the (mouse) cursor" (mouse not necessary)
- line 78f.: I would say "for an asymmetric model the width can be further divided into left and right"
- line 84: "with each other" instead of together
- line 102: "when performing repeated analyses, users familiar with Python may find it helpful to use Fitspy as follows:"
- line 107: "This approach has recently been used by Meyer et al., 2023..."
