+++ paper/paper.md
@@ -21,15 +21,14 @@ bibliography: paper.bib
 # Summary
 
 **`Fitspy`** is a dedicated tool to **fit** **sp**ectra in **py**thon.
-
-Spectrum decomposition (also known as fitting) is the choice of a linear combination of peak models to best represent an experimental spectrum. `Fitspy` currently has implementations for Gaussian, Lorentzian, related asymmetric functions and the Pseudovoigt model. User-defined models can also be added. For the fitting, the model parameters can be subject to bounds and constraints.
+Spectrum decomposition (also known as fitting) is the choice of a linear combination of peak models to best represent an experimental spectrum. `Fitspy` currently has implementations for Gaussian, Lorentzian, related asymmetric functions and the Pseudovoigt as peak models. User-defined models can also be added. For the fitting, the model parameters can be subject to bounds and constraints.
 From a practical point of view, the `Fitspy` GUI (see \autoref{fig:GUI}) is designed to be as simple and intuitive to use as possible.
 
-![The `Fitpsy` GUI.\label{fig:GUI}](https://cea-metrocarac.github.io/fitspy/fitspy.png){width=85%}
+![The `Fitpsy` GUI featuring application options on the right-hand side, and in the central area, the current experimental spectrum displayed in black, the corresponding fit in blue, and the various modeled peaks shown in different colors. \label{fig:GUI}](https://cea-metrocarac.github.io/fitspy/fitspy.png){width=85%}
 
 # Statement of need
 
-The analysis of spectra in many areas of physics, from materials characterisation to astrophysics, often requires their decomposition into more or less complex models to estimate the chemical composition of the subject being analysed.
+The analysis of spectra in many areas of physics, from materials characterisation to astrophysics, often requires their decomposition into more or less complex models to estimate the composition of the subject being analysed.
 To carry out these decompositions, research communities can rely on spectral decomposition tools.
 While many open-source tools for spectral decomposition exist, most of them have, however, been designed for specific application domains, offering a broad range of services beyond mere spectral fitting. Consequently, these tools can prove challenging to use, especially for less experienced individuals.
 
@@ -39,45 +38,45 @@ In the vein of generic tools like `Fityk` [@Fityk] or `PRISMA`  [@PRISMA], `Fits
 
 * **Python Implementation**: `Fitspy` is coded in Python. As a result, for individuals with basic knowledge of the language, spectra can be easily processed by scripts.
 
-* **User Models**: `Fitspy` allows users to input their own models either in the form of literal expressions (for simple models) or through Python scripts (for more complex models).
+* **User Models**: `Fitspy` allows the user to input their own models either in the form of literal expressions (for simple models) or through Python scripts (for more complex models).
 
-* **2D Maps**: `Fitspy` has been designed to handle spectra derived from 2D acquisitions. Note that "2D" can encompass time or any other dimension. When dealing with 2D data, an interactive map in the `Fitspy` GUI allows users to locate and select spectra of interest easily.
+* **2D Maps**: `Fitspy` has been designed to handle spectra derived from 2D acquisitions. Note that "2D" can encompass time or any other dimension. When dealing with 2D data, an interactive map in the `Fitspy` GUI allows the user to locate and select spectra of interest easily.
 
 * **Multiprocessing Capabilities**: `Fitspy` enables spectral fit processing on multiple processors, enhancing efficiency.
 
-* **Constrained Parameters**: Leveraging the `lmfit` library [@lmfit], `Fitspy` empowers users to impose constraints on parameter ranges or establish constraints between parameters using literal expressions.
[…]
