### 2024-04-04 #5868 @Kevin-Mattheus-Moerman
- [ ] Please create a tagged release on your repository (i.e. one that will be listed here: https://github.com/CEA-MetroCarac/fitspy/tags), its version tag should match what we listed here and what is listed for the archive here which is currently `2024.4` (if you end up creating a new version/archive, let us know then we can update the version tag and archive link here to match).
- [ ] In the paper replace `softwares` by `software`  e.g. one could phrase your sentence using `...although other open-source software packages are much more....`
- [ ] Change `visualization`, `characterize`, `customize` (which is American English) to `visualisation`, `characterise`, and `customise`, to be consistent with the rest of the European English used throughout the paper.
