### 2024-04-10 b8213a5
change softwares and visualization, characterize, customize words in European English
