### 2024-04-10 0205abc
add JOSS reviewers and editor acknowledgements
