### 2024-04-18 #5868 @patquem
Thank you all (@maurov , @FCMeng , @phibeck  and @Kevin-Mattheus-Moerman ) for reviewing, editing, and publishing the article.
The **Citations** and the **Acknowlegments** section of the Github FITSPY site have been updated :)
